Data.Validate.Monadic

Description

Documentation

newtype ValidateM e m a

The composition of Validation with another Applicative functor.

Validation is not a Monad, and ValidateM is not a monad transformer. It is equipped with a useful bind function, but that function does not have the right type to make ValidateM a Monad (besides which it would break the laws)

Constructors

ValidateM

Fields

unValidateM :: Compose m (Validation e) a

runValidateM :: ValidateM e m a -> m (Validation e a)

Unwrap a ValidateM

bindVM :: Monad m => m a -> (a -> ValidateM e m b) -> ValidateM e m b

Bind into a ValidateM. Note that the first parameter is m a, not ValidateM e m a.

liftVM0 :: (Functor m, Semigroup e) => m a -> ValidateM e m a

Lift into a succeeding validation

liftVM1 :: (forall x. m x -> m x) -> ValidateM e m a -> ValidateM e m a

Run a natural transformation across ValidateM to alter m

errorVM :: Applicative m => e -> ValidateM e m a

Lift an error into ValidateM

errorVM1 :: (Applicative f, Applicative m) => e -> ValidateM (f e) m a

Lift an error in an Applicative into ValidateM. This is especially useful if you're using list or NonEmpty to collect errors.