Language.Python

Description

hpython provides tools for working with Python source code.

Language.Python.DSL:

A DSL for writing Python programs

Language.Python.Optics:

Optics for working with Python sytnax trees

Language.Python.Parse:

Parse Python source into a syntax tree

Language.Python.Render:

Pretty print Python syntax trees

Language.Python.Syntax:

The data structures that represent Python programs, like Statement and Expr

Language.Python.Validate:

Validate aspects of Python syntax trees, like indentation, syntax, or scope

Documentation

module Language.Python.DSL

module Language.Python.Optics

module Language.Python.Parse

module Language.Python.Render

module Language.Python.Syntax

module Language.Python.Validate