Language.Python.DSL

Description

Passing [] to a function which expects a [Raw Line] is the same as passing [line_ pass_]

Documentation

(&) :: a -> (a -> b) -> b infixl 1

& is a reverse application operator. This provides notational convenience. Its precedence is one higher than that of the forward application operator $, which allows & to be nested in $.

>>> 5 & (+1) & show
"6"

Since: base-4.8.0.0

type Raw f = f '[] ()

Raw represents unvalidated, un-annotated terms.

data Module v a

A Python Module, which is stored as a sequence of statements. A module corresponds to one source file of Python code.

data Statement (v :: [*]) a

A Statement is either a SmallStatement or a CompoundStatement

https://docs.python.org/3.5/reference/compound_stmts.html#compound-statements

data Expr (v :: [*]) a

This large sum type covers all valid Python expressions

Modules

module_ :: [Raw Line] -> Raw Module

Create a Module

>>> module_
>>> [ line_ $ def_ "a" [] [line_ pass_]
>>> , blank_
>>> , line_ $ def_ "b" [] [line_ pass_]
>>> ]
def a():
    pass

def b():
    pass

Identifiers

id_ :: String -> Raw Ident

Ident has an IsString instance, but when a type class dispatches on an Ident we will run into ambiguity if we try to use OverloadedStrings. In these cases we can use id_ to provide the extra type information

data Ident (v :: [*]) a

An identifier. Like many types in hpython, it has an optional annotation and tracks its trailing whitespace.

Raw Idents have an IsString instance.

See https://docs.python.org/3.5/reference/lexical_analysis.html#identifiers

Constructors

MkIdent

Fields

_identAnn :: a

_identValue :: String

_identWhitespace :: [Whitespace]

Lenses

identAnn :: Lens (Ident v a) (Ident v a) a a

identValue :: Lens (Ident v a) (Ident '[] a) String String

identWhitespace :: Lens (Ident v a) (Ident v a) [Whitespace] [Whitespace]

Starred values

class HasStar s t | t -> s where

Minimal complete definition

s_

Methods

s_ :: Raw s -> Raw t

star_ :: Raw Param

Unnamed starred parameter

>>> def_ "a" [ p_ "b", star_ ] [ line_ pass_ ]
def a(b, *):
    pass

Double-starred values

class HasDoubleStar s t | t -> s where

Minimal complete definition

ss_

Methods

ss_ :: Raw s -> Raw t

as syntax

class As s t u | s t -> u, u -> s t where

Minimal complete definition

as_

Methods

as_ :: Raw s -> Raw t -> Raw u

if syntax

class HasIf a where

Minimal complete definition

if_

Methods

if_ :: Raw Expr -> a

for syntax

class HasFor a x | a -> x where

Minimal complete definition

for_

Methods

for_ :: Raw x -> a

in syntax

class HasIn a x | a -> x, x -> a where

Minimal complete definition

in_

Methods

in_ :: Raw Expr -> x -> Raw a infixl 1

data In v a

>>> var_ "a" `in_` var_ "b"
a in b

Constructors

MkIn (Expr v a) (Expr v a)

data InList v a

>>> var_ "a" `in_` [var_ "b", var_ "c"]
a in b, c

Constructors

MkInList (Expr v a) [Expr v a]

: syntax

class HasColon s t | s -> t, t -> s where

Minimal complete definition

(.:)

Methods

(.:) :: Raw s -> Raw Expr -> Raw t infix 0

Comprehensions

comp_ :: Raw e -> Raw CompFor -> [Raw Guard] -> Raw (Comprehension e)

newtype Guard v a

Constructors

MkGuard

Fields

unGuard :: Either (CompFor v a) (CompIf v a)

Parameters and arguments

Parameters

class HasParameters s where

Minimal complete definition

parameters_, parameters

Methods

parameters_ :: Functor f => ([Raw Param] -> f [Raw Param]) -> Raw s -> f (Raw s)

A faux-Lens that allows targeting Params in-between existing formatting, and adding appropriate formatting when extra parameters are introduced.

>>> showStatement myStatement
"def a(b ,  c   ):\n    pass"

>>> showStatement (myStatement '&' '_Fundef' '.' 'parameters_' '.~' ['p_' "d", 'p_' "e"]
"def a(d ,  e   ):\n    pass"

>>> showStatement (myStatement '&' '_Fundef' '.' 'parameters_' '.~' ['p_' "d", 'p_' "e", 'p_' "f"]
"def a(d ,  e   , f):\n    pass"

parameters :: Lens' (Raw s) (CommaSep (Raw Param))

Arguments

class HasArguments s where

Minimal complete definition

setArguments, getArguments

Methods

setArguments :: [Raw Arg] -> Raw s -> Raw s

getArguments :: Raw s -> [Raw Arg]

Positional

class HasPositional p v | p -> v, v -> p where

Positional parameters/arguments

p_ :: Raw Expr -> Raw Arg

p_ :: Raw Ident -> Raw Param

Minimal complete definition

p_

Methods

p_ :: Raw v -> Raw p

data PositionalParam v a

Constructors

MkPositionalParam

Fields

_ppAnn :: a

_ppName :: Ident v a

_ppType :: Maybe (Colon, Expr v a)

_PositionalParam :: Prism (Param v a) (Param '[] a) (PositionalParam v a) (PositionalParam '[] a)

Lenses

ppAnn :: forall v a. Lens' (PositionalParam v a) a

ppName :: forall v a. Lens' (PositionalParam v a) (Ident v a)

ppType :: forall v a. Lens' (PositionalParam v a) (Maybe (Colon, Expr v a))

Keyword

class HasKeyword p where

Keyword parameters/arguments

p_ :: Raw Expr -> Raw Expr -> Raw Arg

p_ :: Raw Ident -> Raw Expr -> Raw Param

Minimal complete definition

k_

Methods

k_ :: Raw Ident -> Raw Expr -> Raw p

data KeywordParam v a

Constructors

MkKeywordParam

Fields

_kpAnn :: a

_kpName :: Ident v a

_kpType :: Maybe (Colon, Expr v a)

_kpEquals :: [Whitespace]

_kpExpr :: Expr v a

_KeywordParam :: Prism (Param v a) (Param '[] a) (KeywordParam v a) (KeywordParam '[] a)

Lenses

kpAnn :: forall v a. Lens' (KeywordParam v a) a

kpName :: forall v a. Lens' (KeywordParam v a) (Ident v a)

kpType :: forall v a. Lens' (KeywordParam v a) (Maybe (Colon, Expr v a))

kpEquals :: forall v a. Lens' (KeywordParam v a) [Whitespace]

kpExpr :: forall v a. Lens' (KeywordParam v a) (Expr v a)

Decorators

decorated_ :: HasDecorators s => [Raw Expr] -> Raw s -> Raw s

class HasDecorators s where

Minimal complete definition

setDecorators, getDecorators, decorators

Methods

setDecorators :: [Raw Expr] -> Raw s -> Raw s

getDecorators :: Raw s -> [Raw Expr]

decorators :: Lens' (Raw s) [Raw Decorator]

Statements

async

class HasAsync s where

Minimal complete definition

async_

Methods

async_ :: Raw s -> Raw s

Lines of code

blank_ :: Raw Line

Create a blank Line

class AsLine s where

Convert some data to a Line

Minimal complete definition

line_

Methods

line_ :: Raw s -> Raw Line

newtype Line v a

One or more lines of Python code

Constructors

Line

Fields

unLine :: Either (Blank a, Newline) (Statement v a)

Block bodies

class HasBody s where

Minimal complete definition

body_, body

Methods

body_ :: Functor f => ([Raw Line] -> f [Raw Line]) -> Raw s -> f (Raw s)

A faux-Lens that targets lines in the body of some statement-piece, but does so 'around' indentation.

>>> def_ "a" [] [ line_ pass_, line_ pass_ ]
def a ():
    pass
    pass

>>> def_ "a" [] [ line_ pass_, line_ pass_ ] ^. body_
pass
pass

>>> def_ "a" [] [ line_ pass_, line_ pass_ ] & body_ .~ [ line_ $ var_ "b" += 1 ]
def a():
    b += 1

>>> def_ "a" [] [ line_ pass_, line_ pass_ ] & body_ <>~ [ line_ $ var_ "b" += 1 ]
def a():
    pass
    pass
    b += 1

body :: Lens' (Raw s) (Raw Suite)

Function definitions

def_ :: Raw Ident -> [Raw Param] -> [Raw Line] -> Raw Statement

>>> def_ "f" [p_ "x"] [line_ $ return_ "x"]
def f(x):
    return x

>>> def_ "f" [p_ "x", k_ "y" 2] [line_ $ return_ "x"]
def f(x, y=2):
    return x

>>> def_ "f" [p_ "x", k_ "y" 2, s_ "z"] [line_ $ return_ "x"]
def f(x, y=2, *z):
    return x

>>> def_ "f" [p_ "x", k_ "y" 2, s_ "z", ss_ "w"] [line_ $ return_ "x"]
def f(x, y=2, *z, **w)
    return x

>>> def_ "f" [p_ "x" .: "String"] [line_ $ return_ "x"]
def f(x: String):
    return x

data Fundef v a

Constructors

MkFundef

Fields

_fdAnn :: a

_fdDecorators :: [Decorator v a]

_fdIndents :: Indents a

_fdAsync :: Maybe (NonEmpty Whitespace)

_fdDefSpaces :: NonEmpty Whitespace

_fdName :: Ident v a

_fdLeftParenSpaces :: [Whitespace]

_fdParameters :: CommaSep (Param v a)

_fdRightParenSpaces :: [Whitespace]

_fdReturnType :: Maybe ([Whitespace], Expr v a)

_fdBody :: Suite v a

_Fundef :: Prism (Statement v a) (Statement '[] a) (Fundef v a) (Fundef '[] a)

mkFundef :: Raw Ident -> [Raw Line] -> Raw Fundef

Create a minimal valid function definition

Lenses

fdAnn :: forall v a. Lens' (Fundef v a) a

fdDecorators :: forall v a. Lens' (Fundef v a) [Decorator v a]

fdIndents :: forall v a. Lens' (Fundef v a) (Indents a)

fdAsync :: forall v a. Lens' (Fundef v a) (Maybe (NonEmpty Whitespace))

fdDefSpaces :: forall v a. Lens' (Fundef v a) (NonEmpty Whitespace)

fdName :: forall v a. Lens' (Fundef v a) (Ident v a)

fdLeftParenSpaces :: forall v a. Lens' (Fundef v a) [Whitespace]

fdParameters :: forall v a. Lens' (Fundef v a) (CommaSep (Param v a))

fdRightParenSpaces :: forall v a. Lens' (Fundef v a) [Whitespace]

fdReturnType :: forall v a. Lens' (Fundef v a) (Maybe ([Whitespace], Expr v a))

fdBody :: forall v a. Lens' (Fundef v a) (Suite v a)

Class definitions

class_ :: Raw Ident -> [Raw Arg] -> [Raw Line] -> Raw Statement

>>> class_ "A" [] [line_ pass_]
class A:
    pass

data ClassDef v a

Constructors

MkClassDef

Fields

_cdAnn :: a

_cdDecorators :: [Decorator v a]

_cdIndents :: Indents a

_cdClass :: NonEmpty Whitespace

_cdName :: Ident v a

_cdArguments :: Maybe ([Whitespace], Maybe (CommaSep1' (Arg v a)), [Whitespace])

_cdBody :: Suite v a

_ClassDef :: Prism (Statement v a) (Statement '[] a) (ClassDef v a) (ClassDef '[] a)

mkClassDef :: Raw Ident -> [Raw Line] -> Raw ClassDef

Create a minimal ClassDef

Lenses

cdAnn :: forall v a. Lens' (ClassDef v a) a

cdDecorators :: forall v a. Lens' (ClassDef v a) [Decorator v a]

cdIndents :: forall v a. Lens' (ClassDef v a) (Indents a)

cdClass :: forall v a. Lens' (ClassDef v a) (NonEmpty Whitespace)

cdName :: forall v a. Lens' (ClassDef v a) (Ident v a)

cdArguments :: forall v a. Lens' (ClassDef v a) (Maybe ([Whitespace], Maybe (CommaSep1' (Arg v a)), [Whitespace]))

cdBody :: forall v a. Lens' (ClassDef v a) (Suite v a)

Assignment

chainEq :: Raw Expr -> [Raw Expr] -> Raw Statement

Chained assignment

>>> chainEq (var_ "a") []
a

>>> chainEq (var_ "a") [var_ "b", var_ "c"]
a = b = c

(.=) :: Raw Expr -> Raw Expr -> Raw Statement infix 0

a = b

(.+=) :: Raw Expr -> Raw Expr -> Raw Statement infix 0

a += b

(.-=) :: Raw Expr -> Raw Expr -> Raw Statement infix 0

a -= b

(.*=) :: Raw Expr -> Raw Expr -> Raw Statement infix 0

a *= b

(.@=) :: Raw Expr -> Raw Expr -> Raw Statement infix 0

a = b@

(./=) :: Raw Expr -> Raw Expr -> Raw Statement infix 0

a /= b

(.%=) :: Raw Expr -> Raw Expr -> Raw Statement infix 0

a %= b

(.&=) :: Raw Expr -> Raw Expr -> Raw Statement infix 0

a &= b

(.|=) :: Raw Expr -> Raw Expr -> Raw Statement infix 0

a |= b

(.^=) :: Raw Expr -> Raw Expr -> Raw Statement infix 0

a ^= b

(.<<=) :: Raw Expr -> Raw Expr -> Raw Statement infix 0

a <<= b

(.>>=) :: Raw Expr -> Raw Expr -> Raw Statement infix 0

a >>= b

(.**=) :: Raw Expr -> Raw Expr -> Raw Statement infix 0

a **= b

(.//=) :: Raw Expr -> Raw Expr -> Raw Statement infix 0

a //= b

Exceptions

tryE_ :: [Raw Line] -> Raw Except -> Raw TryExcept

try ... except with optional else and optional finally

tryF_ :: [Raw Line] -> [Raw Line] -> Raw TryFinally

try:
    ...
finally:
    ...

class HasExcept s where

Minimal complete definition

except_, exceptAs_

Methods

except_ :: [Raw Line] -> s -> Raw TryExcept

exceptAs_ :: AsExceptAs e => Raw e -> [Raw Line] -> s -> Raw TryExcept

You can use exceptAs_ without a binder:

exceptAs_ :: Raw Expr -> [Raw Line] -> Raw s -> Raw TryExcept

exceptAs_ (var_ "Exception") body

or with a binder:

exceptAs_ :: Raw ExceptAs -> [Raw Line] -> Raw s -> Raw TryExcept

exceptAs_ (var_ "Exception" `as_` id_ "a") body

class HasFinally s where

Minimal complete definition

finally_

Methods

finally_ :: [Raw Line] -> Raw s -> Raw s

class AsTry s where

Minimal complete definition

_Try

Methods

_Try :: Prism (Statement v a) (Statement '[] a) (s v a) (s '[] a)

data TryExcept v a

Constructors

MkTryExcept

Fields

_teAnn :: a

_teIndents :: Indents a

_teTry :: [Whitespace]

_teBody :: Suite v a

_teExcepts :: NonEmpty (Except v a)

_teElse :: Maybe (Else v a)

_teFinally :: Maybe (Finally v a)

_TryExcept :: Prism (Statement v a) (Statement '[] a) (TryExcept v a) (TryExcept '[] a)

mkTryExcept :: [Raw Line] -> Raw Except -> Raw TryExcept

Create a minimal valid TryExcept

data TryFinally v a

Constructors

MkTryFinally

Fields

_tfAnn :: a

_tfIndents :: Indents a

_tfTry :: [Whitespace]

_tfBody :: Suite v a

_tfFinally :: Finally v a

_TryFinally :: Prism (Statement v a) (Statement '[] a) (TryFinally v a) (TryFinally '[] a)

mkTryFinally :: [Raw Line] -> [Raw Line] -> Raw TryFinally

Create a minimal valid TryFinally

data ExceptAs (v :: [*]) a

See https://docs.python.org/3.5/reference/compound_stmts.html#the-try-statement

Constructors

ExceptAs

Fields

_exceptAsAnn :: a

_exceptAsExpr :: Expr v a

<expr>

_exceptAsName :: Maybe ([Whitespace], Ident v a)

['as' <ident>]

class AsExceptAs s where

Minimal complete definition

toExceptAs

Methods

toExceptAs :: Raw s -> Raw ExceptAs

data Except v a

Constructors

MkExcept

Fields

_exceptIndents :: Indents a

_exceptExcept :: [Whitespace]

_exceptExceptAs :: Maybe (ExceptAs v a)

_exceptBody :: Suite v a

_Except :: Iso' (Except v a) (Indents a, [Whitespace], Maybe (ExceptAs v a), Suite v a)

mkExcept :: [Raw Line] -> Raw Except

Create a minimal valid Except

data Finally v a

Constructors

MkFinally

Fields

_finallyIndents :: Indents a

_finallyFinally :: [Whitespace]

_finallyBody :: Suite v a

_Finally :: Iso' (Finally v a) (Indents a, [Whitespace], Suite v a)

mkFinally :: [Raw Line] -> Raw Finally

Create a minimal valid Finally

Lenses

teAnn :: forall v a. Lens' (TryExcept v a) a

teIndents :: forall v a. Lens' (TryExcept v a) (Indents a)

teTry :: forall v a. Lens' (TryExcept v a) [Whitespace]

teBody :: forall v a. Lens' (TryExcept v a) (Suite v a)

teExcepts :: forall v a. Lens' (TryExcept v a) (NonEmpty (Except v a))

teElse :: forall v a. Lens' (TryExcept v a) (Maybe (Else v a))

teFinally :: forall v a. Lens' (TryExcept v a) (Maybe (Finally v a))

exceptIndents :: forall v a. Lens' (Except v a) (Indents a)

exceptExcept :: forall v a. Lens' (Except v a) [Whitespace]

exceptExceptAs :: forall v a. Lens' (Except v a) (Maybe (ExceptAs v a))

exceptBody :: forall v a. Lens' (Except v a) (Suite v a)

finallyIndents :: forall v a. Lens' (Finally v a) (Indents a)

finallyFinally :: forall v a. Lens' (Finally v a) [Whitespace]

finallyBody :: forall v a v. Lens (Finally v a) (Finally v a) (Suite v a) (Suite v a)

With statements

with_ :: AsWithItem e => NonEmpty (Raw e) -> [Raw Line] -> Raw Statement

with_ :: NonEmpty (Raw Expr) -> [Raw Line] -> Raw Statement
with_ :: NonEmpty (Raw WithItem) -> [Raw Line] -> Raw Statement

>>> with_ [var_ "a"] [line_ $ var_ "b"]
with a:
    b

>>> with_ [var_ "a" `as_` id_ "name"] [line_ $ var_ "b"]
with a as name:
    b

>>> with_ [withItem_ e Nothing] [line_ $ var_ "b"]
with a:
    b

withItem_ :: Raw Expr -> Maybe (Raw Expr) -> Raw WithItem

data With v a

Constructors

MkWith

Fields

_withAnn :: a

_withIndents :: Indents a

_withAsync :: Maybe (NonEmpty Whitespace)

_withWith :: [Whitespace]

_withItems :: CommaSep1 (WithItem v a)

_withBody :: Suite v a

_With :: Prism (Statement v a) (Statement '[] a) (With v a) (With '[] a)

mkWith :: NonEmpty (Raw WithItem) -> [Raw Line] -> Raw With

Create a minimal valid With

class AsWithItem s where

Minimal complete definition

toWithItem

Methods

toWithItem :: Raw s -> Raw WithItem

data WithItem (v :: [*]) a

See https://docs.python.org/3.5/reference/compound_stmts.html#the-with-statement

Constructors

WithItem

Fields

_withItemAnn :: a

_withItemValue :: Expr v a

<expr>

_withItemBinder :: Maybe ([Whitespace], Expr v a)

['as' <spaces> <expr>]

Lenses

withAnn :: forall v a. Lens' (With v a) a

withIndents :: forall v a. Lens' (With v a) (Indents a)

withAsync :: forall v a. Lens' (With v a) (Maybe (NonEmpty Whitespace))

withWith :: forall v a. Lens' (With v a) [Whitespace]

withItems :: forall v a. Lens' (With v a) (CommaSep1 (WithItem v a))

withBody :: forall v a. Lens' (With v a) (Suite v a)

Flow control

Else clauses

If, While, For, and TryExcept statements can have an Else component.

else_ is considered to be a modifier on these structures.

---

If ... Else:

>>> if_ false_ [line_ pass_] & else_ [line_ pass_]
if False:
    pass
else:
    pass

---

While ... Else:

>>> while_ false_ [line_ pass_] & else_ [line_ pass_]
while False:
    pass
else:
    pass

---

For ... Else:

>>> for_ (var_ "a" `in_` [var_ b]) [line_ pass_] & else_ [line_ pass_]
for a in b:
    pass
else:
    pass

---

TryExcept ... Else:

>>> tryE_ [line_ pass_] & except_ [line_ pass_] & else_ [line_ pass_]
try:
    pass
except:
    pass
else:
    pass

else_ :: HasElse s => [Raw Line] -> Raw s -> Raw s

class HasElse s where

Minimal complete definition

getElse, setElse

Methods

getElse :: Raw s -> Maybe (Raw Else)

setElse :: [Whitespace] -> Maybe (Raw Else) -> Raw s -> Raw s

Break

break_ :: Raw Statement

For loops

For loops are built using for_ syntax:

>>> for_ (var_ "a" `in_` [1, 2, 3]) [line_ (call_ "print" [var_ "a"])]
for a in 1, 2, 3:
    print(a)

See also: HasFor

forSt_ :: Raw Expr -> [Raw Expr] -> [Raw Line] -> Raw For

data For v a

Constructors

MkFor

Fields

_forAnn :: a

_forIndents :: Indents a

_forAsync :: Maybe (NonEmpty Whitespace)

_forFor :: [Whitespace]

_forBinder :: Expr v a

_forIn :: [Whitespace]

_forCollection :: CommaSep1' (Expr v a)

_forBody :: Suite v a

_forElse :: Maybe (Else v a)

_For :: Prism (Statement v a) (Statement '[] a) (For v a) (For '[] a)

mkFor :: Raw Expr -> [Raw Expr] -> [Raw Line] -> Raw For

If statements

ifThen_ :: Raw Expr -> [Raw Line] -> Raw If

elif_ :: Raw Expr -> [Raw Line] -> Raw If -> Raw If

data If v a

Constructors

MkIf

Fields

_ifAnn :: a

_ifIndents :: Indents a

_ifIf :: [Whitespace]

_ifCond :: Expr v a

_ifBody :: Suite v a

_ifElifs :: [Elif v a]

_ifElse :: Maybe (Else v a)

_If :: Prism (Statement v a) (Statement '[] a) (If v a) (If '[] a)

mkIf :: Raw Expr -> [Raw Line] -> Raw If

Create a minimal valid If

data Elif v a

Constructors

MkElif

Fields

_elifIndents :: Indents a

_elifElif :: [Whitespace]

_elifCond :: Expr v a

_elifBody :: Suite v a

_Elif :: Iso' (Elif v a) (Indents a, [Whitespace], Expr v a, Suite v a)

mkElif :: Raw Expr -> [Raw Line] -> Raw Elif

Create a minimal valid Elif

data Else v a

Constructors

MkElse

Fields

_elseIndents :: Indents a

_elseElse :: [Whitespace]

_elseBody :: Suite v a

_Else :: Iso' (Else v a) (Indents a, [Whitespace], Suite v a)

mkElse :: [Raw Line] -> Raw Else

Create a minimal valid Else

Lenses

ifAnn :: forall v a. Lens' (If v a) a

ifIndents :: forall v a. Lens' (If v a) (Indents a)

ifIf :: forall v a. Lens' (If v a) [Whitespace]

ifCond :: forall v a. Lens' (If v a) (Expr v a)

ifBody :: forall v a. Lens' (If v a) (Suite v a)

ifElifs :: forall v a. Lens' (If v a) [Elif v a]

ifElse :: forall v a. Lens' (If v a) (Maybe (Else v a))

elifIndents :: forall v a. Lens' (Elif v a) (Indents a)

elifElif :: forall v a. Lens' (Elif v a) [Whitespace]

elifCond :: forall v a. Lens' (Elif v a) (Expr v a)

elifBody :: forall v a. Lens' (Elif v a) (Suite v a)

elseIndents :: forall v a. Lens' (Else v a) (Indents a)

elseElse :: forall v a. Lens' (Else v a) [Whitespace]

elseBody :: forall v a v. Lens (Else v a) (Else v a) (Suite v a) (Suite v a)

Pass

pass_ :: Raw Statement

>>> pass_
pass

Return

return_ :: Raw Expr -> Raw Statement

>>> return_ (var_ "a")
return a

While loops

while_ :: Raw Expr -> [Raw Line] -> Raw While

data While v a

Constructors

MkWhile

Fields

_whileAnn :: a

_whileIndents :: Indents a

_whileWhile :: [Whitespace]

_whileCond :: Expr v a

_whileBody :: Suite v a

_whileElse :: Maybe (Else v a)

_While :: Prism (Statement v a) (Statement '[] a) (While v a) (While '[] a)

mkWhile :: Raw Expr -> [Raw Line] -> Raw While

Create a minimal valid While

Lenses

whileAnn :: forall v a. Lens' (While v a) a

whileIndents :: forall v a. Lens' (While v a) (Indents a)

whileWhile :: forall v a. Lens' (While v a) [Whitespace]

whileCond :: forall v a. Lens' (While v a) (Expr v a)

whileBody :: forall v a. Lens' (While v a) (Suite v a)

Expressions

expr_ :: Raw Expr -> Raw Statement

Turns an Expr into a Statement

>>> expr_ (int_ 3)
3

var_ :: String -> Raw Expr

await

await_ :: Raw Expr -> Raw Expr

>>> await (var_ "a")
await a

... if ... else ...

ifThenElse_ :: Raw Expr -> Raw Expr -> Raw Expr -> Raw Expr

>>> ifThenElse_ (var_ "a") (var_ "b") (var_ "c")
a if b else c

Generators

gen_ :: Raw (Comprehension Expr) -> Raw Expr

>>> gen_ $ comp_ (var_ "a") (for_ $ var_ "a" `in_` list_ [li_ $ int_ 1, li_ $ int_ 2, li_ $ int_ 3]) [if_ $ var_ "a" .== 2]
(a for a in [1, 2, 3] if a == 2)

yield

yield_ :: [Raw Expr] -> Raw Expr

>>> yield_ []
yield

>>> yield_ [var_ "a"]
yield a

>>> yield_ [var_ "a", var_ "b"]
yield a, b

yield from ...

yieldFrom_ :: Raw Expr -> Raw Expr

>>> yieldFrom_ (var_ "a")
yield from a

Tuples

tuple_ :: [Raw TupleItem] -> Raw Expr

>>> tuple_ []
()

>>> tuple_ [ti_ $ var_ "a"]
a,

>>> tuple_ [s_ $ var_ "a"]
(*a),

>>> tuple_ [ti_ $ var_ "a", ti_ $ var_ "b"]
a, b

>>> tuple_ [ti_ $ var_ "a", s_ $ var_ "b"]
a, *b

data Tuple v a

Constructors

MkTuple

Fields

_tupleAnn :: a

_tupleHead :: TupleItem v a

_tupleComma :: Comma

_tupleTail :: Maybe (CommaSep1' (TupleItem v a))

_Tuple :: Prism (Expr v a) (Expr '[] a) (Tuple v a) (Tuple '[] a)

class AsTupleItem e where

Minimal complete definition

ti_

Methods

ti_ :: Raw e -> Raw TupleItem

Create a TupleItem

data TupleItem (v :: [*]) a

a or *a

Used to construct tuples, e.g. (1, x, **c)

Function calls

call_ :: Raw Expr -> [Raw Arg] -> Raw Expr

>>> call_ "f" [p_ $ var_ "x"]
f(x)

>>> call_ "f" [p_ $ var_ "x", k_ "y" 2]
f(x, y=2)

>>> call_ "f" [p_ $ var_ "x", k_ "y" 2, s_ "z"]
f(x, y=2, *z)

>>> call_ "f" [p_ $ var_ "x", k_ "y" 2, s_ "z", ss_ "w"]
f(x, y=2, *z, **w)

data Call v a

Constructors

MkCall

Fields

_callAnn :: a

_callFunction :: Expr v a

_callLeftParen :: [Whitespace]

_callArguments :: Maybe (CommaSep1' (Arg v a))

_callRightParen :: [Whitespace]

_Call :: Prism (Expr v a) (Expr '[] a) (Call v a) (Call '[] a)

mkCall :: Raw Expr -> Raw Call

Create a minimal valid Call

Lenses

callAnn :: forall v a. Lens' (Call v a) a

callFunction :: forall v a. Lens' (Call v a) (Expr v a)

callLeftParen :: forall v a. Lens' (Call v a) [Whitespace]

callArguments :: forall v a. Lens' (Call v a) (Maybe (CommaSep1' (Arg v a)))

callRightParen :: forall v a. Lens' (Call v a) [Whitespace]

Literals

None

none_ :: Raw Expr

>>> none_
None

data None (v :: [*]) a

Constructors

MkNone

Fields

_noneAnn :: a

_noneWhitespace :: [Whitespace]

_None :: Prism (Expr v a) (Expr '[] a) (None v a) (None '[] a)

Lenses

noneAnn :: forall v a v a. Lens (None v a) (None v a) a a

noneWhitespace :: forall v a v. Lens (None v a) (None v a) [Whitespace] [Whitespace]

Strings

str_ :: String -> Raw Expr

Double-quoted string

>>> str_ "asdf"
"asdf"

str'_ :: String -> Raw Expr

Single-quoted string

>>> str_ "asdf"
'asdf'

longStr_ :: String -> Raw Expr

Long double-quoted string

>>> longStr_ "asdf"
"""asdf"""

longStr'_ :: String -> Raw Expr

Long single-quoted string

>>> longStr'_ "asdf"
'''asdf'''

Integers

int_ :: Integer -> Raw Expr

Raw Expr has a Num instance, but sometimes we need to name integers explicitly

>>> int_ 10
10

Booleans

true_ :: Raw Expr

>>> true_
True

false_ :: Raw Expr

>>> false_
False

Ellipses

ellipsis_ :: Raw Expr

>>> ellipsis_
...

Lists

class AsList s where

>>> list_ [li_ $ var_ "a"]
[a]

>>> list_ [s_ $ var_ "a"]
[*a]

>>> list_ [li_ $ var_ "a", s_ $ var_ "b"]
[a, *b]

>>> list_ $ comp_ (var_ "a") (for_ $ var_ "a" `in_` list_ [li_ $ int_ 1, li_ $ int_ 2, li_ $ int_ 3]) [if_ $ var_ "a" .== 2]
[a for a in [1, 2, 3] if a == 2]

Minimal complete definition

list_

Methods

list_ :: s -> Raw Expr

class AsListItem s where

Minimal complete definition

li_

Methods

li_ :: Raw s -> Raw ListItem

Create a ListItem

data ListItem (v :: [*]) a

a or *a

Used to construct lists, e.g. [ 1, x, **c ]

https://docs.python.org/3/reference/expressions.html#list-displays

Dictionaries

class AsDict s where

>>> dict_ [var_ "a" .: 1]
{a: 1}

>>> dict_ [ss_ $ var_ "a"]
{**a}

>>> dict_ [var_ "a" .: 1, ss_ $ var_ "b"]
{a: 1, **b}

>>> dict_ $ comp_ (var_ "a" .: 1) (for_ $ var_ "a" `in_` list_ [li_ $ int_ 1, li_ $ int_ 2, li_ $ int_ 3]) [if_ $ var_ "a" .== 2]
{a: 1 for a in [1, 2, 3] if a == 2}

Minimal complete definition

dict_

Methods

dict_ :: s -> Raw Expr

data DictItem (v :: [*]) a

a : b or **a

Used to construct dictionaries, e.g. { 1: a, 2: b, **c }

https://docs.python.org/3/reference/expressions.html#dictionary-displays

Sets

class AsSet s where

>>> set_ []
set()

>>> set_ [si_ $ var_ "a"]
{a}

>>> set_ [s_ $ var_ "a"]
{*a}

>>> set_ [si_ $ var_ "a", s_ $ var_ "b"]
{a, *b}

>>> set_ $ comp_ (var_ "a") (for_ $ var_ "a" `in_` set_ [si_ $ int_ 1, si_ $ int_ 2, si_ $ int_ 3]) [if_ $ var_ "a" .== 2]
{a for a in [1, 2, 3] if a == 2}

Minimal complete definition

set_

Methods

set_ :: s -> Raw Expr

class AsSetItem s where

Minimal complete definition

si_

Methods

si_ :: Raw s -> Raw SetItem

Create a SetItem

data SetItem (v :: [*]) a

a or *a

Used to construct sets, e.g. { 1, x, **c }

https://docs.python.org/3/reference/expressions.html#set-displays

Lambdas

lambda_ :: [Raw Param] -> Raw Expr -> Raw Expr

>>> lambda_ [p_ "x"] "x"
lambda x: x

>>> lambda_ [p_ "x", k_ "y" 2] ("x" .+ "y")
lambda x, y=2: x + y

>>> lambda_ [p_ "x", k_ "y" 2, s_ "z"] "a"
lambda x, y=2, *z: a

>>> lambda_ [p_ "x", k_ "y" 2, s_ "z", ss_ "w"] "a"
lambda x, y=2, *z, **w: a

Subscripting

subs_ :: Raw Expr -> Raw Expr -> Raw Expr

>>> subs_ (var_ "a") (int_ 1)
a[1]

>>> subs_ (var_ "a") (tuple_ [ti_ $ int_ 1])
a[1,]

>>> subs_ (var_ "a") (tuple_ [ti_ $ int_ 1, ti_ $ int_ 2])
a[1, 2]

>>> subs_ (var_ "a") (tuple_ [s_ $ var_ "b"])
a[((*b),)]

>>> subs_ (var_ "a") (tuple_ [ti_ $ int_ 1, s_ $ var_ "b"])
a[(1, *b)]

Slicing

sliceF_ :: Raw Expr -> Raw Expr

Slice *from* x

>>> subs_ (var_ "a") (sliceF_ $ int_ 0)
a[1:]

>>> sliceF_ $ int_ 0
slice(1, None, None)

sliceFS_ :: Raw Expr -> Raw Expr -> Raw Expr

Slice *from* x, with *step* y

>>> subs_ (var_ "a") (sliceFS_ (int_ 0) (int_ 2))
a[1::2]

>>> sliceFS_ (int_ 0) (int_ 2)
slice(1, None, 2)

sliceT_ :: Raw Expr -> Raw Expr

Slice To x

>>> subs_ (var_ "a") (sliceT_ $ int_ 10)
a[:10]

>>> sliceT_ $ int_ 10
slice(None, 10, None)

sliceTS_ :: Raw Expr -> Raw Expr -> Raw Expr

Slice To x, with Step y

>>> subs_ (var_ "a") (sliceTS_ (int_ 10) (int_ 2))
a[:10:2]

>>> sliceTS_ (int_ 10) (int_ 2)
slice(None, 10, 2)

sliceFT_ :: Raw Expr -> Raw Expr -> Raw Expr

Slice From x To y

>>> subs_ (var_ "a") (sliceFT_ (int_ 1) (int_ 10))
a[1:10]

>>> sliceFT_ (int_ 1) (int_ 10)
slice(1, 10, None)

sliceFTS_ :: Raw Expr -> Raw Expr -> Raw Expr -> Raw Expr

Slice From x To y, with Step z

>>> subs_ (var_ "a") (sliceFTS_ (int_ 1) (int_ 10) (int_ 2))
a[1:10:2]

>>> sliceFTS_ (int_ 1) (int_ 10) (int_ 2)
slice(1, 10, 2)

sliceS_ :: Raw Expr -> Raw Expr

Slice with *step* x

>>> subs_ (var_ "a") (sliceS_ $ int_ (-1))
a[::-1]

>>> sliceS_ $ int_ (-1)
slice(None, None, -1)

fullSlice_ :: Raw Expr

The slice with no bounds

>>> subs_ (var_ "a") fullSlice_
a[:]

>>> fullSlice_
slice(None, None, None)

slice_ :: Maybe (Raw Expr) -> Maybe (Raw Expr) -> Maybe (Raw Expr) -> Raw Expr

A slice object

Represents a call to a function named slice, with 3 arguments. If an argument is a Nothing then it becomes None, and if the argument is a Just then the contents are extracted.

Dereferencing

(/>) :: Raw Expr -> Raw Ident -> Raw Expr infixl 9

>>> var_ "a" /> var_ "b"
a.b

Unary operators

not_ :: Raw Expr -> Raw Expr

not a

neg_ :: Raw Expr -> Raw Expr

-a

pos_ :: Raw Expr -> Raw Expr

+a

compl_ :: Raw Expr -> Raw Expr

~a

Binary operators

Comparison, bitwise, and arithmetic operators have precedences that are consistent with their Python counterparts. This meansPython expressions can be translated to Haskell with minimal parentheses.

Note: this doesn't apply to unary operators (because Haskell doesn't have unary operators), or the boolean operations and_ and or_ (because we ran out of precedence levels)

Boolean operations

or_ :: Raw Expr -> Raw Expr -> Raw Expr

a or b

Does not have a precedence

and_ :: Raw Expr -> Raw Expr -> Raw Expr

a and b

Does not have a precedence

Comparison operations

is_ :: Raw Expr -> Raw Expr -> Raw Expr infixl 1

a is b

isNot_ :: Raw Expr -> Raw Expr -> Raw Expr infixl 1

a is not b

notIn_ :: Raw Expr -> Raw Expr -> Raw Expr infixl 1

a not in b

(.==) :: Raw Expr -> Raw Expr -> Raw Expr infixl 1

a == b

(.>) :: Raw Expr -> Raw Expr -> Raw Expr infixl 1

a > b

(.>=) :: Raw Expr -> Raw Expr -> Raw Expr infixl 1

a >= b

(.<) :: Raw Expr -> Raw Expr -> Raw Expr infixl 1

a < b

(.<=) :: Raw Expr -> Raw Expr -> Raw Expr infixl 1

a <= b

(.!=) :: Raw Expr -> Raw Expr -> Raw Expr infixl 1

a != b

Bitwise operations

(.|) :: Raw Expr -> Raw Expr -> Raw Expr infixl 2

a | b

(.^) :: Raw Expr -> Raw Expr -> Raw Expr infixl 3

a ^ b

(.&) :: Raw Expr -> Raw Expr -> Raw Expr infixl 4

a & b

(.<<) :: Raw Expr -> Raw Expr -> Raw Expr infixl 5

a << b

(.>>) :: Raw Expr -> Raw Expr -> Raw Expr infixl 5

a >> b

Arithmetic operations

(.-) :: Raw Expr -> Raw Expr -> Raw Expr infixl 6

a - b

(.+) :: Raw Expr -> Raw Expr -> Raw Expr infixl 6

a + b

(.*) :: Raw Expr -> Raw Expr -> Raw Expr infixl 7

a * b

(.@) :: Raw Expr -> Raw Expr -> Raw Expr infixl 7

a @ b

(./) :: Raw Expr -> Raw Expr -> Raw Expr infixl 7

a / b

(.//) :: Raw Expr -> Raw Expr -> Raw Expr infixl 7

a // b

(.%) :: Raw Expr -> Raw Expr -> Raw Expr infixl 7

a % b

(.**) :: Raw Expr -> Raw Expr -> Raw Expr infixr 8

a ** b

Miscellaneous

linesToBlock :: [Raw Line] -> Raw Block

Convert a list of Lines to a Block, without indenting them

blockToLines :: Raw Block -> [Raw Line]