Language.Python.Internal.Lexer

Description

Documentation

tokenizeWithTabs

Arguments

:: (AsLexicalError s Char, AsTabError s SrcInfo, AsIncorrectDedent s SrcInfo)

=> FilePath
File name

-> Text
Input to tokenize

-> Either s [PyToken SrcInfo]

Tokenize an input file, inserting indent/level/dedent tokens in appropriate positions according to the block structure.

Source Information

data SrcInfo

Constructors

SrcInfo

Fields

_srcInfoName :: FilePath

_srcInfoLineStart :: !Int

_srcInfoLineEnd :: !Int

_srcInfoColStart :: !Int

_srcInfoColEnd :: !Int

_srcInfoOffsetStart :: !Int

_srcInfoOffsetEnd :: !Int

initialSrcInfo :: FilePath -> SrcInfo

withSrcInfo :: MonadParsec e s m => m (SrcInfo -> a) -> m a

Errors

class AsLexicalError s t | s -> t where

Minimal complete definition

_LexicalError

Methods

_LexicalError :: Prism' s (NonEmpty SourcePos, Maybe (ErrorItem t), Set (ErrorItem t))

unsafeFromLexicalError :: (HasCallStack, AsLexicalError s t) => ParseError t Void -> s

Convert a concrete ParseError to a value that has an instance of AsLexicalError

This function is partial, because our parser will never use FancyError

class AsTabError s a | s -> a where

Minimal complete definition

_TabError

Methods

_TabError :: Prism' s a

class AsIncorrectDedent s a | s -> a where

Minimal complete definition

_IncorrectDedent

Methods

_IncorrectDedent :: Prism' s a

fromTabError :: (AsTabError s a, AsIncorrectDedent s a) => TabError a -> s

Convert a concrete TabError to a value that has an instance of AsTabError

data TabError a

Constructors

TabError a
Tabs and spaces were used inconsistently

IncorrectDedent a
The dedent at the end of a block doesn't match and preceding indents

e.g.

def a():
    if b:
        pass
    else:
        pass
  pass

The final line will cause an IncorrectDedent error

Miscellaneous

tokenize

Arguments

:: AsLexicalError e Char

=> FilePath
File name

-> Text
Input to tokenize

-> Either e [PyToken SrcInfo]

Convert some input to a sequence of tokens. Indent and dedent tokens are not added (see insertTabs)

insertTabs

Arguments

:: (Semigroup a, AsTabError s a, AsIncorrectDedent s a)

=> a
Initial source annotation

-> [PyToken a]
Token stream

-> Either s [PyToken a]

Insert indent and dedent tokens

https://docs.python.org/3.5/reference/lexical_analysis.html#indentation

Megaparsec re-exports

data ParseError t e

ParseError t e represents a parse error parametrized over the token type t and the custom data e.

Note that the stack of source positions contains current position as its head, and the rest of positions allows to track full sequence of include files with topmost source file at the end of the list.

Semigroup and Monoid instances of the data type allow to merge parse errors from different branches of parsing. When merging two ParseErrors, the longest match is preferred; if positions are the same, custom data sets and collections of message items are combined. Note that fancy errors take precedence over trivial errors in merging.

Since: megaparsec-6.0.0

Constructors

TrivialError (NonEmpty SourcePos) (Maybe (ErrorItem t)) (Set (ErrorItem t))
Trivial errors, generated by Megaparsec's machinery. The data constructor includes the stack of source positions, unexpected token (if any), and expected tokens.

FancyError (NonEmpty SourcePos) (Set (ErrorFancy e))
Fancy, custom errors.