Language.Python.Internal.Syntax.IR

Description

This module only exists as our current best solution to decoupling parts of the concrete syntax from abstract syntax. You won't need to care about its existence and hopefully it will be deleted soon.

Documentation

class AsIRError s a | s -> a where

Minimal complete definition

_InvalidUnpacking

Methods

_InvalidUnpacking :: Prism' s a

data IRError a

Constructors

InvalidUnpacking a
Unpacking ( *value ) was used in an invalid position

fromIRError :: AsIRError s a => IRError a -> s

data SmallStatement a

Constructors

MkSmallStatement (SimpleStatement a) [(Semicolon a, SimpleStatement a)] (Maybe (Semicolon a)) (Maybe (Comment a)) (Maybe Newline)

data Statement a

Constructors

SmallStatement (Indents a) (SmallStatement a)

CompoundStatement (CompoundStatement a)

data CompoundStatement a

Constructors

Fundef

Fields

_csAnn :: a

_unsafeCsFundefDecorators :: [Decorator a]

_csIndents :: Indents a

_unsafeCsFundefAsync :: Maybe (NonEmpty Whitespace)

['async' <spaces>]

_unsafeCsFundefDef :: NonEmpty Whitespace

'def' <spaces>

_unsafeCsFundefName :: Ident '[] a

<ident>

_unsafeCsFundefLeftParen :: [Whitespace]

'(' <spaces>

_unsafeCsFundefParameters :: CommaSep (Param a)

<parameters>

_unsafeCsFundefRightParen :: [Whitespace]

')' <spaces>

_unsafeCsFundefReturnType :: Maybe ([Whitespace], Expr a)

['->' <spaces> <expr>]

_unsafeCsFundefBody :: Suite a

<suite>

If

Fields

_csAnn :: a

_csIndents :: Indents a

_unsafeCsIfIf :: [Whitespace]

'if' <spaces>

_unsafeCsIfCond :: Expr a

<expr>

_unsafeCsIfBody :: Suite a

<suite>

_unsafeCsIfElifs :: [(Indents a, [Whitespace], Expr a, Suite a)]

('elif' <spaces> <expr> <suite>)*

_unsafeCsIfElse :: Maybe (Indents a, [Whitespace], Suite a)

['else' <spaces> <suite>]

While

Fields

_csAnn :: a

_csIndents :: Indents a

_unsafeCsWhileWhile :: [Whitespace]

'while' <spaces>

_unsafeCsWhileCond :: Expr a

<expr>

_unsafeCsWhileBody :: Suite a

<suite>

_unsafeCsWhileElse :: Maybe (Indents a, [Whitespace], Suite a)

['else' <spaces> <suite>]

TryExcept

Fields

_csAnn :: a

_csIndents :: Indents a

_unsafeCsTryExceptTry :: [Whitespace]

'try' <spaces>

_unsafeCsTryExceptBody :: Suite a

<suite>

_unsafeCsTryExceptExcepts :: NonEmpty (Indents a, [Whitespace], Maybe (ExceptAs a), Suite a)

('except' <spaces> <except_as> <suite>)+

_unsafeCsTryExceptElse :: Maybe (Indents a, [Whitespace], Suite a)

['else' <spaces> <suite>]

_unsafeCsTryExceptFinally :: Maybe (Indents a, [Whitespace], Suite a)

['finally' <spaces> <suite>]

TryFinally

Fields

_csAnn :: a

_csIndents :: Indents a

_unsafeCsTryFinallyTry :: [Whitespace]

'try' <spaces>

_unsafeCsTryFinallyTryBody :: Suite a

<suite>

_unsafeCsTryFinallyFinallyIndents :: Indents a

_unsafeCsTryFinallyFinally :: [Whitespace]

'finally' <spaces>

_unsafeCsTryFinallyFinallyBody :: Suite a

<suite>

For

Fields

_csAnn :: a

_csIndents :: Indents a

_unsafeCsForAsync :: Maybe (NonEmpty Whitespace)

['async' <spaces>]

_unsafeCsForFor :: [Whitespace]

'for' <spaces>

_unsafeCsForBinder :: Expr a

<expr>

_unsafeCsForIn :: [Whitespace]

'in' <spaces>

_unsafeCsForCollection :: CommaSep1' (Expr a)

<exprs>

_unsafeCsForBody :: Suite a

<suite>

_unsafeCsForElse :: Maybe (Indents a, [Whitespace], Suite a)

['else' <spaces> <suite>]

ClassDef

Fields

_csAnn :: a

_unsafeCsClassDefDecorators :: [Decorator a]

_csIndents :: Indents a

_unsafeCsClassDefClass :: NonEmpty Whitespace

'class' <spaces>

_unsafeCsClassDefName :: Ident '[] a

<ident>

_unsafeCsClassDefArguments :: Maybe ([Whitespace], Maybe (CommaSep1' (Arg a)), [Whitespace])

['(' <spaces> [<args>] ')' <spaces>]

_unsafeCsClassDefBody :: Suite a

<suite>

With

Fields

_csAnn :: a

_csIndents :: Indents a

_unsafeCsWithAsync :: Maybe (NonEmpty Whitespace)

['async' <spaces>]

_unsafeCsWithWith :: [Whitespace]

'with' <spaces>

_unsafeCsWithItems :: CommaSep1 (WithItem a)

<with_items>

_unsafeCsWithBody :: Suite a

<suite>

data SimpleStatement a

Constructors

Return a [Whitespace] (Maybe (Expr a))

Expr a (Expr a)

Assign a (Expr a) (NonEmpty (Equals, Expr a))

AugAssign a (Expr a) (AugAssign a) (Expr a)

Pass a [Whitespace]

Break a [Whitespace]

Continue a [Whitespace]

Global a (NonEmpty Whitespace) (CommaSep1 (Ident '[] a))

Nonlocal a (NonEmpty Whitespace) (CommaSep1 (Ident '[] a))

Del a [Whitespace] (CommaSep1' (Expr a))

Import a (NonEmpty Whitespace) (CommaSep1 (ImportAs ModuleName '[] a))

From a [Whitespace] (RelativeModuleName '[] a) [Whitespace] (ImportTargets '[] a)

Raise a [Whitespace] (Maybe (Expr a, Maybe ([Whitespace], Expr a)))

Assert a [Whitespace] (Expr a) (Maybe (Comma, Expr a))

data Param a

Constructors

PositionalParam

Fields

_paramAnn :: a

_paramName :: Ident '[] a

_paramType :: Maybe (Colon, Expr a)

KeywordParam

Fields

_paramAnn :: a

_paramName :: Ident '[] a

_paramType :: Maybe (Colon, Expr a)

_unsafeKeywordParamWhitespaceRight :: [Whitespace]

_unsafeKeywordParamExpr :: Expr a

StarParam

Fields

_paramAnn :: a

_unsafeStarParamWhitespace :: [Whitespace]

_unsafeStarParamName :: Ident '[] a

_paramType :: Maybe (Colon, Expr a)

UnnamedStarParam

Fields

_paramAnn :: a

_unsafeUnnamedStarParamWhitespace :: [Whitespace]

DoubleStarParam

Fields

_paramAnn :: a

_unsafeDoubleStarParamWhitespace :: [Whitespace]

_paramName :: Ident '[] a

_paramType :: Maybe (Colon, Expr a)

data CompIf a

Constructors

CompIf a [Whitespace] (Expr a)
'if' any_spaces expr

data CompFor a

Constructors

CompFor a [Whitespace] (Expr a) [Whitespace] (Expr a)
for any_spaces targets 'in' any_spaces expr

data Comprehension e a

Constructors

Comprehension a (e a) (CompFor a) [Either (CompFor a) (CompIf a)]
expr comp_for (comp_for | comp_if)*

data Subscript a

Constructors

SubscriptExpr (Expr a)

SubscriptSlice (Maybe (Expr a)) Colon (Maybe (Expr a)) (Maybe (Colon, Maybe (Expr a)))

data DictItem a

Constructors

DictItem

Fields

_dictItemAnn :: a

_unsafeDictItemKey :: Expr a

_unsafeDictItemColon :: Colon

_unsafeDictItemvalue :: Expr a

DictUnpack

Fields

_dictItemAnn :: a

_unsafeDictItemUnpackWhitespace :: [Whitespace]

_unsafeDictItemUnpackValue :: Expr a

data Arg a

Constructors

PositionalArg

Fields

_argAnn :: a

_argExpr :: Expr a

KeywordArg

Fields

_argAnn :: a

_unsafeKeywordArgName :: Ident '[] a

_unsafeKeywordArgWhitespaceRight :: [Whitespace]

_argExpr :: Expr a

StarArg

Fields

_argAnn :: a

_unsafeStarArgWhitespace :: [Whitespace]

_argExpr :: Expr a

DoubleStarArg

Fields

_argAnn :: a

_unsafeDoubleStarArgWhitespace :: [Whitespace]

_argExpr :: Expr a

data Expr a

Constructors

StarExpr

Fields

_unsafeExprAnn :: a

_unsafeStarExprWhitespace :: [Whitespace]

_unsafeStarExprValue :: Expr a

Unit

Fields

_unsafeExprAnn :: a

_unsafeUnitWhitespaceInner :: [Whitespace]

_unsafeUnitWhitespaceRight :: [Whitespace]

Lambda

Fields

_unsafeExprAnn :: a

_unsafeLambdaWhitespace :: [Whitespace]

_unsafeLambdaArgs :: CommaSep (Param a)

_unsafeLambdaColon :: Colon

_unsafeLambdaBody :: Expr a

Yield

Fields

_unsafeExprAnn :: a

_unsafeYieldWhitespace :: [Whitespace]

_unsafeYieldValue :: CommaSep (Expr a)

YieldFrom

Fields

_unsafeExprAnn :: a

_unsafeYieldWhitespace :: [Whitespace]

_unsafeFromWhitespace :: [Whitespace]

_unsafeYieldFromValue :: Expr a

Ternary

Fields

_unsafeExprAnn :: a

_unsafeTernaryValue :: Expr a

_unsafeTernaryWhitespaceIf :: [Whitespace]

_unsafeTernaryCond :: Expr a

_unsafeTernaryWhitespaceElse :: [Whitespace]

_unsafeTernaryElse :: Expr a

ListComp

Fields

_unsafeExprAnn :: a

_unsafeListCompWhitespaceLeft :: [Whitespace]

_unsafeListCompValue :: Comprehension Expr a

_unsafeListCompWhitespaceRight :: [Whitespace]

List

Fields

_unsafeExprAnn :: a

_unsafeListWhitespaceLeft :: [Whitespace]

_unsafeListValues :: Maybe (CommaSep1' (Expr a))

_unsafeListWhitespaceRight :: [Whitespace]

DictComp

Fields

_unsafeExprAnn :: a

_unsafeDictCompWhitespaceLeft :: [Whitespace]

_unsafeDictCompValue :: Comprehension DictItem a

_unsafeDictCompWhitespaceRight :: [Whitespace]

Dict

Fields

_unsafeExprAnn :: a

_unsafeDictWhitespaceLeft :: [Whitespace]

_unsafeDictValues :: Maybe (CommaSep1' (DictItem a))

_unsafeDictWhitespaceRight :: [Whitespace]

SetComp

Fields

_unsafeExprAnn :: a

_unsafeSetCompWhitespaceLeft :: [Whitespace]

_unsafeSetCompValue :: Comprehension Expr a

_unsafeSetCompWhitespaceRight :: [Whitespace]

Set

Fields

_unsafeExprAnn :: a

_unsafeSetWhitespaceLeft :: [Whitespace]

_unsafeSetValues :: CommaSep1' (Expr a)

_unsafeSetWhitespaceRight :: [Whitespace]

Deref

Fields

_unsafeExprAnn :: a

_unsafeDerefValueLeft :: Expr a

_unsafeDerefWhitespaceLeft :: [Whitespace]

_unsafeDerefValueRight :: Ident '[] a

Subscript

Fields

_unsafeExprAnn :: a

_unsafeSubscriptValueLeft :: Expr a

_unsafeSubscriptWhitespaceLeft :: [Whitespace]

_unsafeSubscriptValueRight :: CommaSep1' (Subscript a)

_unsafeSubscriptWhitespaceRight :: [Whitespace]

Call

Fields

_unsafeExprAnn :: a

_unsafeCallFunction :: Expr a

_unsafeCallWhitespaceLeft :: [Whitespace]

_unsafeCallArguments :: Maybe (CommaSep1' (Arg a))

_unsafeCallWhitespaceRight :: [Whitespace]

None

Fields

_unsafeExprAnn :: a

_unsafeNoneWhitespace :: [Whitespace]

Ellipsis

Fields

_unsafeExprAnn :: a

_unsafeEllipsisWhitespace :: [Whitespace]

BinOp

Fields

_unsafeExprAnn :: a

_unsafeBinOpExprLeft :: Expr a

_unsafeBinOpOp :: BinOp a

_unsafeBinOpExprRight :: Expr a

UnOp

Fields

_unsafeExprAnn :: a

_unsafeUnOpOp :: UnOp a

_unsafeUnOpValue :: Expr a

Parens

Fields

_unsafeExprAnn :: a

_unsafeParensWhitespaceLeft :: [Whitespace]

_unsafeParensValue :: Expr a

_unsafeParensWhitespaceAfter :: [Whitespace]

Ident

Fields

_unsafeIdentValue :: Ident '[] a

Int

Fields

_unsafeExprAnn :: a

_unsafeIntValue :: IntLiteral a

_unsafeIntWhitespace :: [Whitespace]

Float

Fields

_unsafeExprAnn :: a

_unsafeFloatValue :: FloatLiteral a

_unsafeFloatWhitespace :: [Whitespace]

Imag

Fields

_unsafeExprAnn :: a

_unsafeImagValue :: ImagLiteral a

_unsafeImagWhitespace :: [Whitespace]

Bool

Fields

_unsafeExprAnn :: a

_unsafeBoolValue :: Bool

_unsafeBoolWhitespace :: [Whitespace]

String

Fields

_unsafeExprAnn :: a

_unsafeStringLiteralValue :: NonEmpty (StringLiteral a)

Tuple

Fields

_unsafeExprAnn :: a

_unsafeTupleHead :: Expr a

_unsafeTupleWhitespace :: Comma

_unsafeTupleTail :: Maybe (CommaSep1' (Expr a))

Not

Fields

_unsafeExprAnn :: a

_unsafeNotWhitespace :: [Whitespace]

_unsafeNotValue :: Expr a

Generator

Fields

_unsafeExprAnn :: a

_generatorValue :: Comprehension Expr a

Await

Fields

_unsafeExprAnn :: a

_unsafeAwaitWhitespace :: [Whitespace]

_unsafeAwaitValue :: Expr a

exprAnn :: Lens' (Expr a) a

data Suite a

Constructors

SuiteOne a Colon (SmallStatement a)

SuiteMany a Colon (Maybe (Comment a)) Newline (Block a)

data Block a

Constructors

Block

Fields

_blockBlankLines :: [(Blank a, Newline)]

_blockHead :: Statement a

_blockTail :: [Either (Blank a, Newline) (Statement a)]

data WithItem a

Constructors

WithItem

Fields

_withItemAnn :: a

_withItemValue :: Expr a

_withItemBinder :: Maybe ([Whitespace], Expr a)

data Decorator a

Constructors

Decorator

Fields

_decoratorAnn :: a

_decoratorIndents :: Indents a

_decoratorAt :: At

_decoratorExpr :: Expr a

_decoratorComment :: Maybe (Comment a)

_decoratorNewline :: Newline

_decoratorBlankLines :: [(Blank a, Newline)]

data ExceptAs a

Constructors

ExceptAs

Fields

_exceptAsAnn :: a

_exceptAsExpr :: Expr a

_exceptAsName :: Maybe ([Whitespace], Ident '[] a)

data Module a

Constructors

ModuleEmpty

ModuleBlankFinal (Blank a)

ModuleBlank (Blank a) Newline (Module a)

ModuleStatement (Statement a) (Module a)

data FromIRContext

Constructors

FromIRContext

Fields

_allowStarred :: Bool

allowStarred :: Iso' FromIRContext Bool

fromIR_expr :: AsIRError e a => Expr a -> Validation (NonEmpty e) (Expr '[] a)

fromIR_suite :: AsIRError e a => Suite a -> Validation (NonEmpty e) (Suite '[] a)

fromIR_param :: AsIRError e a => Param a -> Validation (NonEmpty e) (Param '[] a)

fromIR_arg :: AsIRError e a => Arg a -> Validation (NonEmpty e) (Arg '[] a)

fromIR_decorator :: AsIRError e a => Decorator a -> Validation (NonEmpty e) (Decorator '[] a)

fromIR_exceptAs :: AsIRError e a => ExceptAs a -> Validation (NonEmpty e) (ExceptAs '[] a)

fromIR_withItem :: AsIRError e a => WithItem a -> Validation (NonEmpty e) (WithItem '[] a)

fromIR_comprehension :: AsIRError e a => (ex a -> Validation (NonEmpty e) (ex' '[] a)) -> Comprehension ex a -> Validation (NonEmpty e) (Comprehension ex' '[] a)

fromIR_dictItem :: AsIRError e a => DictItem a -> Validation (NonEmpty e) (DictItem '[] a)

fromIR_subscript :: AsIRError e a => Subscript a -> Validation (NonEmpty e) (Subscript '[] a)

fromIR_block :: AsIRError e a => Block a -> Validation (NonEmpty e) (Block '[] a)

fromIR_compFor :: AsIRError e a => CompFor a -> Validation (NonEmpty e) (CompFor '[] a)

fromIR_compIf :: AsIRError e a => CompIf a -> Validation (NonEmpty e) (CompIf '[] a)

fromIR_smallStatement :: AsIRError e a => SmallStatement a -> Validation (NonEmpty e) (SmallStatement '[] a)

fromIR_statement :: AsIRError e a => Statement a -> Validation (NonEmpty e) (Statement '[] a)

fromIR_SimpleStatement :: AsIRError e a => SimpleStatement a -> Validation (NonEmpty e) (SimpleStatement '[] a)

fromIR_compoundStatement :: AsIRError e a => CompoundStatement a -> Validation (NonEmpty e) (CompoundStatement '[] a)

fromIR_listItem :: AsIRError e a => Expr a -> Validation (NonEmpty e) (ListItem '[] a)

fromIR_tupleItem :: AsIRError e a => Expr a -> Validation (NonEmpty e) (TupleItem '[] a)

fromIR_setItem :: AsIRError e a => Expr a -> Validation (NonEmpty e) (SetItem '[] a)

fromIR :: AsIRError e a => Module a -> Validation (NonEmpty e) (Module '[] a)