Language.Python.Internal.Syntax.Strings.Raw

Description

This module provides a datatype for Python's "raw strings" - strings that are prefixed with an r. Raw strings are specified here:

https://docs.python.org/3.5/reference/lexical_analysis.html#string-and-bytes-literals

In spite of Python's noble intentions to treat backslashes as first-class citizens, the lexer still occasionally lets its latent character-ism out.

In particular, a raw string cannot end in an odd number of backslashes.

r"\\" means "a string consisting of two backslash characters", but there is no way to represent "a string consisting of three backslash characters" using raw string syntax.

Thus, a RawString is "a sequence of of Chars ending in an even number of backslash characters"

Documentation

data RawString s

_RawString :: (AsEmpty s, Cons s s Char Char) => Prism' s (RawString s)