Language.Python.Optics

Description

Optics for manipulating Python syntax trees

Documentation

module Language.Python.Optics.Validated

Indentation

class HasIndents s a | s -> a where

Minimal complete definition

_Indents

Methods

_Indents :: Traversal' s (Indents a)

Traversal' targeting the indentation inside a structure

Note: whitespace inside 'enclosed forms' (such as lists or tuples) is not considered indentation.

e.g.

In the following code, there is only one chunk of indentation:

def a():
    [ b
    , c
    , d
    ]

it's here:

def a():
    [ b
^^^^
    , c
    , d
    ]

The rest is whitespace that is internal to the list.

_Indent :: HasIndents s a => Traversal' s [Whitespace]

Traversal' targeting the indent-chunks in a structure

e.g.

This is one indent chunk:

def a():
    pass
    if b:
        pass
^^^^

and this is another

def a():
    pass
    if b:
        pass
    ^^^^

Newlines

class HasNewlines s where

Minimal complete definition

_Newlines

Methods

_Newlines :: Traversal' (s v a) Newline

Traversal' targeting all of thie Newlines in a structure

Simple statements

Assignment

assignTargets :: Traversal (Expr v a) (Expr '[] a) (Ident v a) (Ident '[] a)

Compound statements

Function defintions

_Fundef :: Prism (Statement v a) (Statement '[] a) (Fundef v a) (Fundef '[] a)

Class defintions

_ClassDef :: Prism (Statement v a) (Statement '[] a) (ClassDef v a) (ClassDef '[] a)

while statements

_While :: Prism (Statement v a) (Statement '[] a) (While v a) (While '[] a)

for statements

_For :: Prism (Statement v a) (Statement '[] a) (For v a) (For '[] a)

with statements

_With :: Prism (Statement v a) (Statement '[] a) (With v a) (With '[] a)

if statements

_If :: Prism (Statement v a) (Statement '[] a) (If v a) (If '[] a)

_Elif :: Iso' (Elif v a) (Indents a, [Whitespace], Expr v a, Suite v a)

try statements

_TryExcept :: Prism (Statement v a) (Statement '[] a) (TryExcept v a) (TryExcept '[] a)

_TryFinally :: Prism (Statement v a) (Statement '[] a) (TryFinally v a) (TryFinally '[] a)

_Finally :: Iso' (Finally v a) (Indents a, [Whitespace], Suite v a)

_Except :: Iso' (Except v a) (Indents a, [Whitespace], Maybe (ExceptAs v a), Suite v a)

class AsTry s where

Minimal complete definition

_Try

Methods

_Try :: Prism (Statement v a) (Statement '[] a) (s v a) (s '[] a)

else

_Else :: Iso' (Else v a) (Indents a, [Whitespace], Suite v a)

Parameters

_PositionalParam :: Prism (Param v a) (Param '[] a) (PositionalParam v a) (PositionalParam '[] a)

_KeywordParam :: Prism (Param v a) (Param '[] a) (KeywordParam v a) (KeywordParam '[] a)

_UnnamedStarParam :: Prism (Param v a) (Param '[] a) (UnnamedStarParam v a) (UnnamedStarParam '[] a)

_StarParam :: Prism (Param v a) (Param '[] a) (StarParam v a) (StarParam '[] a)

Expressions

Identifiers

_Ident :: Prism (Expr v a) (Expr '[] a) (Ident v a) (Ident '[] a)

None

_None :: Prism (Expr v a) (Expr '[] a) (None v a) (None '[] a)

Function calls

_Call :: Prism (Expr v a) (Expr '[] a) (Call v a) (Call '[] a)

Tuples

_Tuple :: Prism (Expr v a) (Expr '[] a) (Tuple v a) (Tuple '[] a)

_TupleUnpack :: Prism (TupleItem v a) (TupleItem '[] a) (TupleUnpack v a) (TupleUnpack '[] a)

tupleItems :: Traversal (Tuple v a) (Tuple '[] a) (TupleItem v a) (TupleItem '[] a)

Lists

_List :: Prism (Expr v a) (Expr '[] a) (List v a) (List '[] a)

_ListUnpack :: Prism (ListItem v a) (ListItem '[] a) (ListUnpack v a) (ListUnpack '[] a)

listItems :: Traversal (List v a) (List '[] a) (ListItem v a) (ListItem '[] a)