Language.Python.Optics.Indents

Description

Documentation

_Indent :: HasIndents s a => Traversal' s [Whitespace]

Traversal' targeting the indent-chunks in a structure

e.g.

This is one indent chunk:

def a():
    pass
    if b:
        pass
^^^^

and this is another

def a():
    pass
    if b:
        pass
    ^^^^

class HasIndents s a | s -> a where

Minimal complete definition

_Indents

Methods

_Indents :: Traversal' s (Indents a)

Traversal' targeting the indentation inside a structure

Note: whitespace inside 'enclosed forms' (such as lists or tuples) is not considered indentation.

e.g.

In the following code, there is only one chunk of indentation:

def a():
    [ b
    , c
    , d
    ]

it's here:

def a():
    [ b
^^^^
    , c
    , d
    ]

The rest is whitespace that is internal to the list.