Language.Python.Optics.Newlines

Description

Documentation

class HasNewlines s where

Minimal complete definition

_Newlines

Methods

_Newlines :: Traversal' s Newline

Traversal' targeting all of thie Newlines in a structure

This only targets places that contain the Newline datatype; it doesn't target newline characters in string literals, for example.