Language.Python.Optics.Validated

Description

Documentation

class Validated (s :: [*] -> * -> *) where

A type class for things for which we can strip the validation information. This can help types line up when they need to, for example to put many things of various validation statuses together in a list.

Methods

unvalidated :: Getter (s v a) (s '[] a)

unvalidated :: Coercible (s v a) (s '[] a) => Getter (s v a) (s '[] a)