Language.Python.Parse.Error

Documentation

data ParseError a

Constructors

LexicalError (NonEmpty SourcePos) (Maybe (ErrorItem Char)) (Set (ErrorItem Char))
An error occured during tokenization (this is a re-packed megaparsec error)

ParseError (NonEmpty SourcePos) (Maybe (ErrorItem (PyToken a))) (Set (ErrorItem (PyToken a)))
An error occured during parsing (this is a re-packed megaparsec error)

TabError a
Tabs and spaces were used inconsistently

IncorrectDedent a
The dedent at the end of a block doesn't match and preceding indents

e.g.

def a():
    if b:
        pass
    else:
        pass
  pass

The final line will cause an IncorrectDedent error

InvalidUnpacking a
Unpacking ( *value ) was used in an invalid position

Classy Prisms

class AsLexicalError s t | s -> t where

Minimal complete definition

_LexicalError

Methods

_LexicalError :: Prism' s (NonEmpty SourcePos, Maybe (ErrorItem t), Set (ErrorItem t))

class AsTabError s a | s -> a where

Minimal complete definition

_TabError

Methods

_TabError :: Prism' s a

class AsIncorrectDedent s a | s -> a where

Minimal complete definition

_IncorrectDedent

Methods

_IncorrectDedent :: Prism' s a

class AsIRError s a | s -> a where

Minimal complete definition

_InvalidUnpacking

Methods

_InvalidUnpacking :: Prism' s a

class AsParseError s t | s -> t where

Minimal complete definition

_ParseError

Methods

_ParseError :: Prism' s (NonEmpty SourcePos, Maybe (ErrorItem t), Set (ErrorItem t))

Megaparsec re-exports

data ErrorItem t

Data type that is used to represent “unexpected/expected” items in ParseError. The data type is parametrized over the token type t.

Since: megaparsec-5.0.0

Constructors

Tokens (NonEmpty t)
Non-empty stream of tokens

Label (NonEmpty Char)
Label (cannot be empty)

EndOfInput
End of input

data SourcePos

The data type SourcePos represents source positions. It contains the name of the source file, a line number, and a column number. Source line and column positions change intensively during parsing, so we need to make them strict to avoid memory leaks.

Constructors

SourcePos

Fields

sourceName :: FilePath

Name of source file

sourceLine :: !Pos

Line number

sourceColumn :: !Pos

Column number