Language.Python.Syntax

Documentation

type Raw f = f '[] ()

data Statement (v :: [*]) a

data Expr (v :: [*]) a

Parameters and arguments

Parameters

class HasParameters s where

Minimal complete definition

setParameters, parameters

Methods

setParameters :: [Raw Param] -> Raw s -> Raw s

parameters :: Lens' (Raw s) (CommaSep (Raw Param))

Positional

class HasPositional p v | p -> v where

Positional parameters/arguments

p_ :: Raw Expr -> Raw Arg

p_ :: Raw Ident -> Raw Param

Minimal complete definition

p_

Methods

p_ :: v -> p

Keyword

class HasKeyword p where

Keyword parameters/arguments

Minimal complete definition

k_

Methods

k_ :: Raw Ident -> Raw Expr -> p

data KeywordParam v a

Constructors

MkKeywordParam

Fields

_kpAnn :: a

_kpName :: Ident v a

_kpType :: Maybe ([Whitespace], Expr v a)

_kpEquals :: [Whitespace]

_kpExpr :: Expr v a

_KeywordParam :: Prism (Param v a) (Param '[] a) (KeywordParam v a) (KeywordParam '[] a)

Lenses

kpAnn :: forall v a. Lens' (KeywordParam v a) a

kpName :: forall v a. Lens' (KeywordParam v a) (Ident v a)

kpType :: forall v a. Lens' (KeywordParam v a) (Maybe ([Whitespace], Expr v a))

kpEquals :: forall v a. Lens' (KeywordParam v a) [Whitespace]

kpExpr :: forall v a. Lens' (KeywordParam v a) (Expr v a)

Decorators

class HasDecorators s where

Minimal complete definition

setDecorators, getDecorators, decorators

Methods

setDecorators :: [Raw Expr] -> Raw s -> Raw s

getDecorators :: Raw s -> [Raw Expr]

decorators :: Lens' (Raw s) [Raw Decorator]

decorated_ :: HasDecorators s => [Raw Expr] -> Raw s -> Raw s

Statements

Block bodies

blank_ :: Raw Line

st_ :: Raw Statement -> Raw Line

newtype Line v a

Constructors

Line

Fields

unLine :: Either ([Whitespace], Newline) (Statement v a)

class HasBody s where

forall ws s. id ~ getBody . flip (setBody ws) s

forall ws. let f x = setBody ws (getBody x) x in f ~ f . f

Minimal complete definition

setBody, getBody, body

Methods

setBody

Arguments

:: [Whitespace]
Indentation scheme for the new lines

-> NonEmpty (Raw Line)
Lines to become the new body

-> Raw s
Current code

-> Raw s

Set the block body of some code, first indententing the block by a specified amount, then indenting it by the level of its parent

For example, if we wanted to set the body of function_b to

while True:
  pass

in

def function_a():
  def function_b():
    pass
    if True:
      pass
    else:
      pass

then we would get

def function_a():
  def function_b():
    while True:
      pass

getBody :: Raw s -> NonEmpty (Raw Line)

View the block body of some code without the indentation from its outer context

For example, if we ran getBody function_b in this code:

def function_a():
  def function_b():
    pass
    if True:
      pass
    else:
      pass

then we would get

pass
if True:
  pass
else:
  pass

body :: Lens' (Raw s) (Raw Suite)

modifyBody

Arguments

:: HasBody s

=> [Whitespace]
New indentation scheme

-> (NonEmpty (Raw Line) -> NonEmpty (Raw Line))
Modification function

-> Raw s
Existing code

-> Raw s

Modify the block body of some code

Function definitions

data Fundef v a

Constructors

Fundef

Fields

_fdAnn :: a

_fdDecorators :: [Decorator v a]

_fdIndents :: Indents a

_fdAsync :: Maybe (NonEmpty Whitespace)

_fdDefSpaces :: NonEmpty Whitespace

_fdName :: Ident v a

_fdLeftParenSpaces :: [Whitespace]

_fdParameters :: CommaSep (Param v a)

_fdRightParenSpaces :: [Whitespace]

_fdReturnType :: Maybe ([Whitespace], Expr v a)

_fdBody :: Suite v a

_Fundef :: Prism (Statement v a) (Statement '[] a) (Fundef v a) (Fundef '[] a)

mkFundef :: Raw Ident -> NonEmpty (Raw Line) -> Raw Fundef

def_ :: Raw Ident -> [Raw Param] -> NonEmpty (Raw Line) -> Raw Statement

Lenses

fdAnn :: forall v a. Lens' (Fundef v a) a

fdDecorators :: forall v a. Lens' (Fundef v a) [Decorator v a]

fdIndents :: forall v a. Lens' (Fundef v a) (Indents a)

fdAsync :: forall v a. Lens' (Fundef v a) (Maybe (NonEmpty Whitespace))

fdDefSpaces :: forall v a. Lens' (Fundef v a) (NonEmpty Whitespace)

fdName :: forall v a. Lens' (Fundef v a) (Ident v a)

fdLeftParenSpaces :: forall v a. Lens' (Fundef v a) [Whitespace]

fdParameters :: forall v a. Lens' (Fundef v a) (CommaSep (Param v a))

fdRightParenSpaces :: forall v a. Lens' (Fundef v a) [Whitespace]

fdReturnType :: forall v a. Lens' (Fundef v a) (Maybe ([Whitespace], Expr v a))

fdBody :: forall v a. Lens' (Fundef v a) (Suite v a)

Assignment

(.=) :: Raw Expr -> Raw Expr -> Raw Statement

Flow control

break_ :: Raw Statement

for_ :: Raw Expr -> Raw Expr -> NonEmpty (Raw Line) -> Raw Statement

if_ :: Raw Expr -> NonEmpty (Raw Line) -> Raw Statement

ifElse_ :: Raw Expr -> NonEmpty (Raw Line) -> NonEmpty (Raw Line) -> Raw Statement

pass_ :: Raw Statement

return_ :: Raw Expr -> Raw Statement

while_ :: Raw Expr -> NonEmpty (Raw Line) -> Raw Statement

Expressions

expr_ :: Raw Expr -> Raw Statement

var_ :: String -> Raw Expr

call_ :: Raw Expr -> [Raw Arg] -> Raw Expr

Literals

list_ :: [Raw Expr] -> Raw Expr

none_ :: Raw Expr

str_ :: String -> Raw Expr

true_ :: Raw Expr

false_ :: Raw Expr

Dereferencing

(/>) :: Raw Expr -> Raw Ident -> Raw Expr infixl 9

Binary operators

is_ :: Raw Expr -> Raw Expr -> Raw Expr infixl 1

(.*) :: Raw Expr -> Raw Expr -> Raw Expr infixl 7

(.-) :: Raw Expr -> Raw Expr -> Raw Expr infixl 6

(.+) :: Raw Expr -> Raw Expr -> Raw Expr infixl 6

(.==) :: Raw Expr -> Raw Expr -> Raw Expr infixl 1