Language.Python.Syntax.CommaSep

Description

Documentation

newtype Comma

The venerable comma separator

Constructors

MkComma [Whitespace]

data CommaSep a

Items separated by commas, with optional whitespace following each comma

Constructors

CommaSepNone

CommaSepOne a

CommaSepMany a Comma (CommaSep a)

_CommaSep :: Iso (Maybe (a, [(Comma, a)], Maybe Comma)) (Maybe (b, [(Comma, b)], Maybe Comma)) (CommaSep a) (CommaSep b)

Iso to unpack a CommaSep

csTrailingWhitespace :: HasTrailingWhitespace a => Traversal' (CommaSep a) [Whitespace]

Traversal targeting the trailing whitespace in a comma separated list.

This can't be an instance of HasTrailingWhitespace because CommaSepNone never has trailing whitespace.

appendCommaSep :: [Whitespace] -> CommaSep a -> CommaSep a -> CommaSep a

Appends two comma separated values together.

The provided whitespace is to follow the joining comma which is added

maybeToCommaSep :: Maybe a -> CommaSep a

Convert a maybe to a singleton or nullary CommaSep

listToCommaSep :: [a] -> CommaSep a

Convert a list to a CommaSep

Anywhere where whitespace is ambiguous, this function puts a single space

data CommaSep1 a

Non-empty CommaSep

Constructors

CommaSepOne1 a

CommaSepMany1 a Comma (CommaSep1 a)

commaSep1Head :: CommaSep1 a -> a

Get the first element of a CommaSep1

appendCommaSep1 :: [Whitespace] -> CommaSep1 a -> CommaSep1 a -> CommaSep1 a

Appends two non-empty comma separated values together.

The provided whitespace is to follow the joining comma which is added

listToCommaSep1 :: NonEmpty a -> CommaSep1 a

Convert a NonEmpty to a CommaSep1

Anywhere where whitespace is ambiguous, this function puts a single space

listToCommaSep1' :: [a] -> Maybe (CommaSep1' a)

Attempt to insert comma separators into a list, which will not be terminated by a comma.

If the list is empty, Nothing is returned.

data CommaSep1' a

Non-empty CommaSep, optionally terminated by a comma

Assumes that the contents consumes trailing whitespace

Constructors

CommaSepOne1' a (Maybe Comma)

CommaSepMany1' a Comma (CommaSep1' a)

_CommaSep1' :: Iso (a, [(Comma, a)], Maybe Comma) (b, [(Comma, b)], Maybe Comma) (CommaSep1' a) (CommaSep1' b)

Iso to unpack a CommaSep1'