Language.Python.Syntax.Expr

Description

Expressions

data Expr (v :: [*]) a

This large sum type covers all valid Python expressions

Constructors

Unit

Fields

_unsafeExprAnn :: a

_unsafeUnitWhitespaceInner :: [Whitespace]

_unsafeUnitWhitespaceRight :: [Whitespace]

Lambda

Fields

_unsafeExprAnn :: a

_unsafeLambdaWhitespace :: [Whitespace]

_unsafeLambdaArgs :: CommaSep (Param v a)

_unsafeLambdaColon :: Colon

_unsafeLambdaBody :: Expr v a

Yield

Fields

_unsafeExprAnn :: a

_unsafeYieldWhitespace :: [Whitespace]

_unsafeYieldValue :: Maybe (Expr v a)

YieldFrom

Fields

_unsafeExprAnn :: a

_unsafeYieldWhitespace :: [Whitespace]

_unsafeFromWhitespace :: [Whitespace]

_unsafeYieldFromValue :: Expr v a

Ternary

Fields

_unsafeExprAnn :: a

_unsafeTernaryValue :: Expr v a

_unsafeTernaryWhitespaceIf :: [Whitespace]

_unsafeTernaryCond :: Expr v a

_unsafeTernaryWhitespaceElse :: [Whitespace]

_unsafeTernaryElse :: Expr v a

ListComp

Fields

_unsafeExprAnn :: a

_unsafeListCompWhitespaceLeft :: [Whitespace]

_unsafeListCompValue :: Comprehension Expr v a

_unsafeListCompWhitespaceRight :: [Whitespace]

List

Fields

_unsafeExprAnn :: a

_unsafeListWhitespaceLeft :: [Whitespace]

_unsafeListValues :: Maybe (CommaSep1' (ListItem v a))

_unsafeListWhitespaceRight :: [Whitespace]

DictComp

Fields

_unsafeExprAnn :: a

_unsafeDictCompWhitespaceLeft :: [Whitespace]

_unsafeDictCompValue :: Comprehension DictItem v a

_unsafeDictCompWhitespaceRight :: [Whitespace]

Dict

Fields

_unsafeExprAnn :: a

_unsafeDictWhitespaceLeft :: [Whitespace]

_unsafeDictValues :: Maybe (CommaSep1' (DictItem v a))

_unsafeDictWhitespaceRight :: [Whitespace]

SetComp

Fields

_unsafeExprAnn :: a

_unsafeSetCompWhitespaceLeft :: [Whitespace]

_unsafeSetCompValue :: Comprehension SetItem v a

_unsafeSetCompWhitespaceRight :: [Whitespace]

Set

Fields

_unsafeExprAnn :: a

_unsafeSetWhitespaceLeft :: [Whitespace]

_unsafeSetValues :: CommaSep1' (SetItem v a)

_unsafeSetWhitespaceRight :: [Whitespace]

Deref

Fields

_unsafeExprAnn :: a

_unsafeDerefValueLeft :: Expr v a

_unsafeDerefWhitespaceLeft :: [Whitespace]

_unsafeDerefValueRight :: Ident v a

Subscript

Fields

_unsafeExprAnn :: a

_unsafeSubscriptValueLeft :: Expr v a

_unsafeSubscriptWhitespaceLeft :: [Whitespace]

_unsafeSubscriptValueRight :: CommaSep1' (Subscript v a)

_unsafeSubscriptWhitespaceRight :: [Whitespace]

Call

Fields

_unsafeExprAnn :: a

_unsafeCallFunction :: Expr v a

_unsafeCallWhitespaceLeft :: [Whitespace]

_unsafeCallArguments :: Maybe (CommaSep1' (Arg v a))

_unsafeCallWhitespaceRight :: [Whitespace]

None

Fields

_unsafeExprAnn :: a

_unsafeNoneWhitespace :: [Whitespace]

Ellipsis

Fields

_unsafeExprAnn :: a

_unsafeEllipsisWhitespace :: [Whitespace]

BinOp

Fields

_unsafeExprAnn :: a

_unsafeBinOpExprLeft :: Expr v a

_unsafeBinOpOp :: BinOp a

_unsafeBinOpExprRight :: Expr v a

UnOp

Fields

_unsafeExprAnn :: a

_unsafeUnOpOp :: UnOp a

_unsafeUnOpValue :: Expr v a

Parens

Fields

_unsafeExprAnn :: a

_unsafeParensWhitespaceLeft :: [Whitespace]

_unsafeParensValue :: Expr v a

_unsafeParensWhitespaceAfter :: [Whitespace]

Ident

Fields

_unsafeIdentValue :: Ident v a

Int

Fields

_unsafeExprAnn :: a

_unsafeIntValue :: IntLiteral a

_unsafeIntWhitespace :: [Whitespace]

Float

Fields

_unsafeExprAnn :: a

_unsafeFloatValue :: FloatLiteral a

_unsafeFloatWhitespace :: [Whitespace]

Imag

Fields

_unsafeExprAnn :: a

_unsafeImagValue :: ImagLiteral a

_unsafeImagWhitespace :: [Whitespace]

Bool

Fields

_unsafeExprAnn :: a

_unsafeBoolValue :: Bool

_unsafeBoolWhitespace :: [Whitespace]

String

Fields

_unsafeExprAnn :: a

_unsafeStringValue :: NonEmpty (StringLiteral a)

Tuple

Fields

_unsafeExprAnn :: a

_unsafeTupleHead :: TupleItem v a

_unsafeTupleWhitespace :: Comma

_unsafeTupleTail :: Maybe (CommaSep1' (TupleItem v a))

Not

Fields

_unsafeExprAnn :: a

_unsafeNotWhitespace :: [Whitespace]

_unsafeNotValue :: Expr v a

Generator

Fields

_unsafeExprAnn :: a

_generatorValue :: Comprehension Expr v a

Await

Fields

_unsafeExprAnn :: a

_unsafeAwaitWhitespace :: [Whitespace]

_unsafeAwaitValue :: Expr v a

class HasExprs s where

Traversal over all the expressions in a term

Minimal complete definition

_Exprs

Methods

_Exprs :: Traversal (s v a) (s '[] a) (Expr v a) (Expr '[] a)

exprAnn :: Lens' (Expr v a) a

Lens on the top-level annotation in an expression

shouldGroupLeft :: BinOp a -> Expr v a -> Bool

shouldGroupLeft op left returns true if left needs to be parenthesised when it is the left argument of op

shouldGroupRight :: BinOp a -> Expr v a -> Bool

shouldGroupRight op right returns true if right needs to be parenthesised when it is the right argument of op

Parameters and arguments

data Param (v :: [*]) a

Formal parameters for functions

See https://docs.python.org/3.5/reference/compound_stmts.html#function-definitions

Constructors

PositionalParam
def foo(a):

Fields

_paramAnn :: a

_paramName :: Ident v a

_paramType :: Maybe (Colon, Expr v a)

KeywordParam
def foo(bar=None):

Fields

_paramAnn :: a

_paramName :: Ident v a

_paramType :: Maybe (Colon, Expr v a)

_unsafeKeywordParamWhitespaceRight :: [Whitespace]

_unsafeKeywordParamExpr :: Expr v a

StarParam
def foo(*xs):

Fields

_paramAnn :: a

_unsafeStarParamWhitespace :: [Whitespace]

_unsafeStarParamName :: Maybe (Ident v a)

_paramType :: Maybe (Colon, Expr v a)

DoubleStarParam
def foo(**dict):

Fields

_paramAnn :: a

_unsafeDoubleStarParamWhitespace :: [Whitespace]

_paramName :: Ident v a

_paramType :: Maybe (Colon, Expr v a)

paramAnn :: Lens' (Param v a) a

Lens on the syntrax tree annotation on a parameter

paramType :: Lens (Param v a) (Param '[] a) (Maybe (Colon, Expr v a)) (Maybe (Colon, Expr '[] a))

Lens on the optional Python type annotation which may follow a parameter

This lens, like many others in hpython, loses validation information (the v type parameter)

The following is an example, where int is the paramtype:

def foo(x: int):

paramName :: Traversal (Param v a) (Param '[] a) (Ident v a) (Ident '[] a)

(affine) Traversal on the name of a parameter

The name is x in the following examples:

def foo(x):
def foo(x=None):
def foo(*x):
def foo(**x):

But the following example does not have a paramName:

def foo(*):

data Arg (v :: [*]) a

Actual parameters for functions

In the following examples, x is an actual parameter.

y = foo(x)
y = bar(quux=x)
y = baz(*x)
y = flux(**x)

Constructors

PositionalArg

Fields

_argAnn :: a

_argExpr :: Expr v a

KeywordArg

Fields

_argAnn :: a

_unsafeKeywordArgName :: Ident v a

_unsafeKeywordArgWhitespaceRight :: [Whitespace]

_argExpr :: Expr v a

StarArg

Fields

_argAnn :: a

_unsafeStarArgWhitespace :: [Whitespace]

_argExpr :: Expr v a

DoubleStarArg

Fields

_argAnn :: a

_unsafeDoubleStarArgWhitespace :: [Whitespace]

_argExpr :: Expr v a

argExpr :: Lens (Arg v a) (Arg '[] a) (Expr v a) (Expr '[] a)

Lens on the Python expression which is passed as the argument

Comprehension expressions

data Comprehension e (v :: [*]) a

A Python for comprehension, such as

x for y in z

Constructors

Comprehension a (e v a) (CompFor v a) [Either (CompFor v a) (CompIf v a)]
expr comp_for (comp_for | comp_if)*

data CompIf (v :: [*]) a

A condition inside a comprehension, e.g. [x for x in xs if even(x)]

Constructors

CompIf a [Whitespace] (Expr v a)

data CompFor (v :: [*]) a

A nested comprehesion, e.g. [(x, y) for x in xs for y in ys]

Constructors

CompFor a [Whitespace] (Expr v a) [Whitespace] (Expr v a)

Collection items

data DictItem (v :: [*]) a

a : b or **a

Used to construct dictionaries, e.g. { 1: a, 2: b, **c }

Constructors

DictItem

Fields

_dictItemAnn :: a

_unsafeDictItemKey :: Expr v a

_unsafeDictItemColon :: Colon

_unsafeDictItemValue :: Expr v a

DictUnpack

Fields

_dictItemAnn :: a

_unsafeDictItemUnpackWhitespace :: [Whitespace]

_unsafeDictItemUnpackValue :: Expr v a

data ListItem (v :: [*]) a

a or *a

Used to construct lists, e.g. [ 1, x, **c ]

Constructors

ListItem

Fields

_listItemAnn :: a

_unsafeListItemValue :: Expr v a

ListUnpack

Fields

_listItemAnn :: a

_unsafeListUnpackParens :: [([Whitespace], [Whitespace])]

_unsafeListUnpackWhitespace :: [Whitespace]

_unsafeListUnpackValue :: Expr v a

data SetItem (v :: [*]) a

a or *a

Used to construct sets, e.g. { 1, x, **c }

Constructors

SetItem

Fields

_setItemAnn :: a

_unsafeSetItemValue :: Expr v a

SetUnpack

Fields

_setItemAnn :: a

_unsafeSetUnpackParens :: [([Whitespace], [Whitespace])]

_unsafeSetUnpackWhitespace :: [Whitespace]

_unsafeSetUnpackValue :: Expr v a

data TupleItem (v :: [*]) a

a or *a

Used to construct tuples, e.g. (1, x, **c)

Constructors

TupleItem

Fields

_tupleItemAnn :: a

_unsafeTupleItemValue :: Expr v a

TupleUnpack

Fields

_tupleItemAnn :: a

_unsafeTupleUnpackParens :: [([Whitespace], [Whitespace])]

_unsafeTupleUnpackWhitespace :: [Whitespace]

_unsafeTupleUnpackValue :: Expr v a

Subscripts

data Subscript (v :: [*]) a

Syntax for things that can be used as subscripts (inside the square brackets)

e.g. a[b], a[:], a[b:], a[:b], a[b:c], a[b:c:d]

Constructors

SubscriptExpr (Expr v a)

SubscriptSlice (Maybe (Expr v a)) Colon (Maybe (Expr v a)) (Maybe (Colon, Maybe (Expr v a)))