Language.Python.Syntax.Ident

Description

Documentation

data Ident (v :: [*]) a

An identifier. Like many types in hpython, it has an optional annotation and tracks its trailing whitespace.

Raw Idents have an IsString instance.

See https://docs.python.org/3.5/reference/lexical_analysis.html#identifiers

Constructors

MkIdent

Fields

_identAnn :: Ann a

_identValue :: String

_identWhitespace :: [Whitespace]

Lenses

identAnn :: Lens (Ident v a) (Ident v a) a a

identValue :: Lens (Ident v a) (Ident '[] a) String String

identWhitespace :: Lens (Ident v a) (Ident v a) [Whitespace] [Whitespace]

Extra functions

isIdentifierStart :: Char -> Bool

Determine whether this character could start a valid identifier

isIdentifierChar :: Char -> Bool

Determine whether this character could be part of a valid identifier