Language.Python.Syntax.Module

Description

Documentation

data Module v a

A Python Module, which is stored as a sequence of statements. A module corresponds to one source file of Python code.

Constructors

ModuleEmpty

ModuleBlankFinal (Blank a)

ModuleBlank (Blank a) Newline (Module v a)

ModuleStatement (Statement v a) (Module v a)