Language.Python.Syntax.Numbers

Description

Numerical literal values in Python

Datatypes

data IntLiteral a

An integer literal value.

5 is an integer literal.

6.2 is a literal but is not an integer

x might be an integer, but is not a literal

See https://docs.python.org/3.5/reference/lexical_analysis.html#integer-literals

Constructors

IntLiteralDec
Decimal

1234

Fields

_intLiteralAnn :: a

_unsafeIntLiteralDecValue :: NonEmpty DecDigit

IntLiteralBin
Binary

0b10110

Fields

_intLiteralAnn :: a

_unsafeIntLiteralBinUppercase :: Bool

_unsafeIntLiteralBinValue :: NonEmpty BinDigit

IntLiteralOct
Octal

0o1367

Fields

_intLiteralAnn :: a

_unsafeIntLiteralOctUppercase :: Bool

_unsafeIntLiteralOctValue :: NonEmpty OctDigit

IntLiteralHex
Mixed-case hexadecimal

0x18B4f

Fields

_intLiteralAnn :: a

_unsafeIntLiteralHexUppercase :: Bool

_unsafeIntLiteralHexValue :: NonEmpty HeXDigit

data Sign

Positive or negative, as in -7

Constructors

Pos

Neg

data E

When a floating point literal is in scientific notation, it includes the character e, which can be lower or upper case.

Constructors

Ee

EE

data FloatExponent

The exponent of a floating point literal.

An e, followed by an optional Sign, followed by at least one digit.

Constructors

FloatExponent E (Maybe Sign) (NonEmpty DecDigit)

data FloatLiteral a

A literal floating point value.

Eg. 7.63

See https://docs.python.org/3.5/reference/lexical_analysis.html#floating-point-literals

Constructors

FloatLiteralFull
'Complete' floats

12.

12.34

12.e34

12.34e56

Fields

_floatLiteralAnn :: a

_floatLiteralFullLeft :: NonEmpty DecDigit

_floatLiteralFullRight :: Maybe (These (NonEmpty DecDigit) FloatExponent)

FloatLiteralPoint
Floats that begin with a decimal point

.12

.12e34

Fields

_floatLiteralAnn :: a

_floatLiteralPointRight :: NonEmpty DecDigit

_floatLiteralPointExponent :: Maybe FloatExponent

FloatLiteralWhole
Floats with no decimal points

12e34

Fields

_floatLiteralAnn :: a

_floatLiteralWholeRight :: NonEmpty DecDigit

_floatLiteralWholeExponent :: FloatExponent

data ImagLiteral a

Imaginary number literals

See https://docs.python.org/3.5/reference/lexical_analysis.html#imaginary-literals

Constructors

ImagLiteralInt
A decimal integer followed by a 'j'

12j

Fields

_imagLiteralAnn :: a

_unsafeImagLiteralIntValue :: NonEmpty DecDigit

_imagLiteralUppercase :: Bool

ImagLiteralFloat
A float followed by a 'j'

12.j

12.3j

.3j

Fields

_imagLiteralAnn :: a

_unsafeImagLiteralFloatValue :: FloatLiteral a

_imagLiteralUppercase :: Bool

Rendering

The output of these functions is guaranteed to be valid Python code

showIntLiteral :: IntLiteral a -> Text

showFloatLiteral :: FloatLiteral a -> Text

showFloatExponent :: FloatExponent -> Text

showImagLiteral :: ImagLiteral a -> Text