Language.Python.Syntax.Operator.Binary

Description

This module contains a datatype for binary operators and a precedence table with associated operations. This presentation of operators is simpler and more flexible than hard-coding them into the syntax tree.

Documentation

data BinOp a

A Python binary operator, such as +, along with its trailing Whitespace

The type variable allows annotations, but it can simply be made () for an unannotated BinOp.

Constructors

Is a [Whitespace]
a is b

IsNot a [Whitespace] [Whitespace]
a is not b

In a [Whitespace]
a in b

NotIn a [Whitespace] [Whitespace]
a not in b

Minus a [Whitespace]
a - b

Exp a [Whitespace]
a ** b

BoolAnd a [Whitespace]
a and b

BoolOr a [Whitespace]
a or b

Eq a [Whitespace]
a == b

Lt a [Whitespace]
a < b

LtEq a [Whitespace]
a <= b

Gt a [Whitespace]
a > b

GtEq a [Whitespace]
a >= b

NotEq a [Whitespace]
a != b

Multiply a [Whitespace]
a * b

Divide a [Whitespace]
a / b

FloorDivide a [Whitespace]
a // b

Percent a [Whitespace]
a % b

Plus a [Whitespace]
a + b

BitOr a [Whitespace]
a | b

BitXor a [Whitespace]
a ^ b

BitAnd a [Whitespace]
a & b

ShiftLeft a [Whitespace]
a << b

ShiftRight a [Whitespace]
a >> b

At a [Whitespace]
a b@

data Assoc

The associativity of an operator. Each operator is either left-associative or right associative.

Left associative:

x + y + z = (x + y) + z

Right associative:

x + y + z = x + (y + z)

Constructors

L

R

data OpEntry

An operator along with its precedence and associativity.

Constructors

OpEntry

Fields

_opOperator :: BinOp ()

_opPrec :: Int

_opAssoc :: Assoc

opPrec :: Lens' OpEntry Int

opOperator :: Lens' OpEntry (BinOp ())

opAssoc :: Lens' OpEntry Assoc

operatorTable :: [OpEntry]

operatorTable is a list of all operators in ascending order of precedence.

sameOperator :: BinOp a -> BinOp a' -> Bool

Compare two BinOps to determine whether they represent the same operator, ignoring annotations and trailing whitespace.

isComparison :: BinOp a -> Bool

Is a BinOp a comparison, such as <=

lookupOpEntry :: BinOp a -> [OpEntry] -> OpEntry

Retrieve the information for a given operator from the operator table.