Language.Python.Syntax.Operator.Unary

Description

Unary operators

Documentation

data UnOp a

An UnOp is a unary operator in Python, such as - for negation. An operator is stored with an annotation and its trailing whitespace.

Constructors

Negate a [Whitespace]
-a

Positive a [Whitespace]
+a

Complement a [Whitespace]
~a