Language.Python.Syntax.Statement

Description

Statements

data Statement (v :: [*]) a

A Statement is either a SmallStatement or a CompoundStatement

https://docs.python.org/3.5/reference/compound_stmts.html#compound-statements

Constructors

SmallStatement (Indents a) (SmallStatement v a)

CompoundStatement (CompoundStatement v a)

Traversals

class HasStatements s where

Traversal over all the Statements in a term

Minimal complete definition

_Statements

Methods

_Statements :: Traversal (s v a) (s '[] a) (Statement v a) (Statement '[] a)

Decorators

data Decorator (v :: [*]) a

Decorators on function definitions

https://docs.python.org/3.5/reference/compound_stmts.html#function-definitions

https://docs.python.org/3.5/glossary.html#term-decorator

Constructors

Decorator

Fields

_decoratorAnn :: Ann a

_decoratorIndents :: Indents a

Preceding indentation

_decoratorAt :: At

'@' <spaces>

_decoratorExpr :: Expr v a

<expr>

_decoratorComment :: Maybe (Comment a)

Trailing comment

_decoratorNewline :: Newline

Trailing newline

_decoratorBlankLines :: [(Blank a, Newline)]

Trailing blank lines

Compound statements

data CompoundStatement (v :: [*]) a

See https://docs.python.org/3.5/reference/compound_stmts.html

Constructors

Fundef
https://docs.python.org/3.5/reference/compound_stmts.html#function-definitions

https://docs.python.org/3.5/reference/compound_stmts.html#coroutine-function-definition

Fields

_csAnn :: Ann a

_unsafeCsFundefDecorators :: [Decorator v a]

Preceding Decorators

_csIndents :: Indents a

Preceding indentation

_unsafeCsFundefAsync :: Maybe (NonEmpty Whitespace)

['async' <spaces>]

_unsafeCsFundefDef :: NonEmpty Whitespace

'def' <spaces>

_unsafeCsFundefName :: Ident v a

<ident>

_unsafeCsFundefLeftParen :: [Whitespace]

'(' <spaces>

_unsafeCsFundefParameters :: CommaSep (Param v a)

<parameters>

_unsafeCsFundefRightParen :: [Whitespace]

')' <spaces>

_unsafeCsFundefReturnType :: Maybe ([Whitespace], Expr v a)

['->' <spaces> <expr>]

_unsafeCsFundefBody :: Suite v a

<suite>

If
https://docs.python.org/3.5/reference/compound_stmts.html#the-if-statement

Fields

_csAnn :: Ann a

_csIndents :: Indents a

Preceding indentation

_unsafeCsIfIf :: [Whitespace]

'if' <spaces>

_unsafeCsIfCond :: Expr v a

<expr>

_unsafeCsIfBody :: Suite v a

<suite>

_unsafeCsIfElifs :: [(Indents a, [Whitespace], Expr v a, Suite v a)]

('elif' <spaces> <expr> <suite>)*

_unsafeCsIfElse :: Maybe (Indents a, [Whitespace], Suite v a)

['else' <spaces> <suite>]

While
https://docs.python.org/3.5/reference/compound_stmts.html#the-while-statement

Fields

_csAnn :: Ann a

_csIndents :: Indents a

Preceding indentation

_unsafeCsWhileWhile :: [Whitespace]

'while' <spaces>

_unsafeCsWhileCond :: Expr v a

<expr>

_unsafeCsWhileBody :: Suite v a

<suite>

_unsafeCsWhileElse :: Maybe (Indents a, [Whitespace], Suite v a)

['else' <spaces> <suite>]

TryExcept
https://docs.python.org/3.5/reference/compound_stmts.html#the-try-statement

Fields

_csAnn :: Ann a

_csIndents :: Indents a

Preceding indentation

_unsafeCsTryExceptTry :: [Whitespace]

'try' <spaces>

_unsafeCsTryExceptBody :: Suite v a

<suite>

_unsafeCsTryExceptExcepts :: NonEmpty (Indents a, [Whitespace], Maybe (ExceptAs v a), Suite v a)

('except' <spaces> <except_as> <suite>)+

_unsafeCsTryExceptElse :: Maybe (Indents a, [Whitespace], Suite v a)

['else' <spaces> <suite>]

_unsafeCsTryExceptFinally :: Maybe (Indents a, [Whitespace], Suite v a)

['finally' <spaces> <suite>]

TryFinally
https://docs.python.org/3.5/reference/compound_stmts.html#the-try-statement

Fields

_csAnn :: Ann a

_csIndents :: Indents a

Preceding indentation

_unsafeCsTryFinallyTry :: [Whitespace]

'try' <spaces>

_unsafeCsTryFinallyTryBody :: Suite v a

<suite>

_unsafeCsTryFinallyFinallyIndents :: Indents a

_unsafeCsTryFinallyFinally :: [Whitespace]

'finally' <spaces>

_unsafeCsTryFinallyFinallyBody :: Suite v a

<suite>

For
https://docs.python.org/3.5/reference/compound_stmts.html#the-for-statement

https://docs.python.org/3.5/reference/compound_stmts.html#the-async-for-statement

Fields

_csAnn :: Ann a

_csIndents :: Indents a

Preceding indentation

_unsafeCsForAsync :: Maybe (NonEmpty Whitespace)

['async' <spaces>]

_unsafeCsForFor :: [Whitespace]

'for' <spaces>

_unsafeCsForBinder :: Expr v a

<expr>

_unsafeCsForIn :: [Whitespace]

'in' <spaces>

_unsafeCsForCollection :: CommaSep1' (Expr v a)

<exprs>

_unsafeCsForBody :: Suite v a

<suite>

_unsafeCsForElse :: Maybe (Indents a, [Whitespace], Suite v a)

['else' <spaces> <suite>]

ClassDef
https://docs.python.org/3.5/reference/compound_stmts.html#class-definitions

Fields

_csAnn :: Ann a

_unsafeCsClassDefDecorators :: [Decorator v a]

Preceding Decorators

_csIndents :: Indents a

Preceding indentation

_unsafeCsClassDefClass :: NonEmpty Whitespace

'class' <spaces>

_unsafeCsClassDefName :: Ident v a

<ident>

_unsafeCsClassDefArguments :: Maybe ([Whitespace], Maybe (CommaSep1' (Arg v a)), [Whitespace])

['(' <spaces> [<args>] ')' <spaces>]

_unsafeCsClassDefBody :: Suite v a

<suite>

With
https://docs.python.org/3.5/reference/compound_stmts.html#the-with-statement

https://docs.python.org/3.5/reference/compound_stmts.html#the-async-with-statement

Fields

_csAnn :: Ann a

_csIndents :: Indents a

Preceding indentation

_unsafeCsWithAsync :: Maybe (NonEmpty Whitespace)

['async' <spaces>]

_unsafeCsWithWith :: [Whitespace]

'with' <spaces>

_unsafeCsWithItems :: CommaSep1 (WithItem v a)

<with_items>

_unsafeCsWithBody :: Suite v a

<suite>

Small statements

data SmallStatement (v :: [*]) a

See simpl_stmt at https://docs.python.org/3.5/reference/grammar.html. The grammar has the terminology mixed up - it should really be called small_stmt there.

Constructors

MkSmallStatement (SimpleStatement v a) [(Semicolon a, SimpleStatement v a)] (Maybe (Semicolon a)) (Maybe (Comment a)) (Maybe Newline)

Simple statements

data SimpleStatement (v :: [*]) a

https://docs.python.org/3.5/reference/simple_stmts.html

Constructors

Return (Ann a) [Whitespace] (Maybe (Expr v a))
'return' <spaces> [<expr>]

https://docs.python.org/3.5/reference/simple_stmts.html#the-return-statement

Expr (Ann a) (Expr v a)
<expr>

https://docs.python.org/3.5/reference/simple_stmts.html#expression-statements

Assign (Ann a) (Expr v a) (NonEmpty (Equals, Expr v a))
<expr> ('=' <spaces> <expr>)+

https://docs.python.org/3.5/reference/simple_stmts.html#assignment-statements

AugAssign (Ann a) (Expr v a) (AugAssign a) (Expr v a)
<expr> <augassign> <expr>

https://docs.python.org/3.5/reference/simple_stmts.html#augmented-assignment-statements

Pass (Ann a) [Whitespace]
'pass' <spaces>

https://docs.python.org/3.5/reference/simple_stmts.html#the-pass-statement

Break (Ann a) [Whitespace]
'break' <spaces>

https://docs.python.org/3.5/reference/simple_stmts.html#the-break-statement

Continue (Ann a) [Whitespace]
'continue' <spaces>

https://docs.python.org/3.5/reference/simple_stmts.html#the-continue-statement

Global (Ann a) (NonEmpty Whitespace) (CommaSep1 (Ident v a))
'global' <spaces> <idents>

https://docs.python.org/3.5/reference/simple_stmts.html#the-global-statement

Nonlocal (Ann a) (NonEmpty Whitespace) (CommaSep1 (Ident v a))
'nonlocal' <spaces> <idents>

https://docs.python.org/3.5/reference/simple_stmts.html#the-nonlocal-statement

Del (Ann a) [Whitespace] (CommaSep1' (Expr v a))
'del' <spaces> <exprs>

https://docs.python.org/3.5/reference/simple_stmts.html#the-del-statement

Import (Ann a) (NonEmpty Whitespace) (CommaSep1 (ImportAs ModuleName v a))
'import' <spaces> <modulenames>

https://docs.python.org/3.5/reference/simple_stmts.html#the-import-statement

From (Ann a) [Whitespace] (RelativeModuleName v a) [Whitespace] (ImportTargets v a)
'from' <spaces> <relative_module_name> 'import' <spaces> <import_targets>

https://docs.python.org/3.5/reference/simple_stmts.html#the-import-statement

Raise (Ann a) [Whitespace] (Maybe (Expr v a, Maybe ([Whitespace], Expr v a)))
'raise' <spaces> [<expr> ['as' <spaces> <expr>]]

https://docs.python.org/3.5/reference/simple_stmts.html#the-raise-statement

Assert (Ann a) [Whitespace] (Expr v a) (Maybe (Comma, Expr v a))
'assert' <spaces> <expr> [',' <spaces> <expr>]

https://docs.python.org/3.5/reference/simple_stmts.html#the-assert-statement

with ... as ...

data WithItem (v :: [*]) a

See https://docs.python.org/3.5/reference/compound_stmts.html#the-with-statement

Constructors

WithItem

Fields

_withItemAnn :: Ann a

_withItemValue :: Expr v a

<expr>

_withItemBinder :: Maybe ([Whitespace], Expr v a)

['as' <spaces> <expr>]

Lenses

withItemAnn :: forall v a. Lens' (WithItem v a) (Ann a)

withItemValue :: forall v a. Lens' (WithItem v a) (Expr v a)

withItemBinder :: forall v a. Lens' (WithItem v a) (Maybe ([Whitespace], Expr v a))

except ... as ...

data ExceptAs (v :: [*]) a

See https://docs.python.org/3.5/reference/compound_stmts.html#the-try-statement

Constructors

ExceptAs

Fields

_exceptAsAnn :: Ann a

_exceptAsExpr :: Expr v a

<expr>

_exceptAsName :: Maybe ([Whitespace], Ident v a)

['as' <ident>]

Lenses

exceptAsAnn :: forall v a. Lens' (ExceptAs v a) (Ann a)

exceptAsExpr :: forall v a. Lens' (ExceptAs v a) (Expr v a)

exceptAsName :: forall v a. Lens' (ExceptAs v a) (Maybe ([Whitespace], Ident v a))

Suites

data Suite (v :: [*]) a

A compound statement consists of one or more clauses. A clause consists of a header and a suite.

Constructors

SuiteOne (Ann a) Colon (SmallStatement v a)

SuiteMany (Ann a) Colon (Maybe (Comment a)) Newline (Block v a)

Blocks

data Block (v :: [*]) a

A Block is an indented multi-line chunk of code, forming part of a Suite.

Constructors

Block

Fields

_blockBlankLines :: [(Blank a, Newline)]

Blank lines at the beginning of the block

_blockHead :: Statement v a

The first statement of the block

_blockTail :: [Either (Blank a, Newline) (Statement v a)]

The remaining items of the block, which may be statements or blank lines

Lenses

blockBlankLines :: forall v a. Lens' (Block v a) [(Blank a, Newline)]

blockHead :: forall v a. Lens' (Block v a) (Statement v a)

blockTail :: forall v a. Lens' (Block v a) [Either (Blank a, Newline) (Statement v a)]

Traversals

class HasBlocks s where

Minimal complete definition

_Blocks

Methods

_Blocks :: Traversal (s v a) (s '[] a) (Block v a) (Block '[] a)

Traversal targeting all the Blocks in a structure