Language.Python.Syntax.Strings

Description

Python string literals.

See https://docs.python.org/3.5/reference/lexical_analysis.html#string-and-bytes-literals

Datatypes

Characters

data PyChar

A character in a string literal. This is a large sum type, with a catch-all of a Haskell Char.

Constructors

Char_newline
\newline

Char_octal1 OctDigit
\1

Char_octal2 OctDigit OctDigit
\12

Char_octal3 OctDigit OctDigit OctDigit
\123

Char_hex HeXDigit HeXDigit
\xFb

Char_uni16 HeXDigit HeXDigit HeXDigit HeXDigit
\u12aD

Char_uni32 HeXDigit HeXDigit HeXDigit HeXDigit HeXDigit HeXDigit HeXDigit HeXDigit
\Udeadbeef

Char_esc_bslash
\\

Char_esc_singlequote
\'

Char_esc_doublequote
\"

Char_esc_a
\a

Char_esc_b
\b

Char_esc_f
\f

Char_esc_n
\n

Char_esc_r
\r

Char_esc_t
\t

Char_esc_v
\v

Char_lit Char
Any character

fromHaskellString :: String -> [PyChar]

Convert a Haskell string to a list of PyChar. This is useful when writing Python in Haskell.

String information

data QuoteType

Double or single quotation marks?

"Double quotes"
"""Double quotes"""
'Single quotes'
'''Single quotes'''

Constructors

SingleQuote

DoubleQuote

data StringType

Three pairs of quotations or one?

"""Long string"""
'''Also long string'''
"Short string"
'Also short string'

Constructors

ShortString

LongString

String prefixes

data StringPrefix

In Python 3.5, a prefix of u or U is allowed, but doesn't have any meaning. They exist for backwards compatibility with Python 2.

See https://www.python.org/dev/peps/pep-0414/

Constructors

Prefix_u

Prefix_U

data RawStringPrefix

Raw strings are prefixed with either r or R.

Constructors

Prefix_r

Prefix_R

data BytesPrefix

This prefix indicates it's a bytes literal rather than a string literal.

Constructors

Prefix_b

Prefix_B

data RawBytesPrefix

A string of raw bytes can be indicated by a number of prefixes

Constructors

Prefix_br

Prefix_Br

Prefix_bR

Prefix_BR

Prefix_rb

Prefix_rB

Prefix_Rb

Prefix_RB

hasPrefix :: StringLiteral a -> Bool

Most types of StringLiteral have prefixes. Plain old strings may have an optional prefix, but it is meaningless.

String literals

data StringLiteral a

A StringLiteral, complete with a prefix, information about quote type and number, and a list of PyChars.

Like many other data types in hpython, it has an annotation and trailing whitespace.

Constructors

RawStringLiteral

Fields

_stringLiteralAnn :: Ann a

_unsafeRawStringLiteralPrefix :: RawStringPrefix

_stringLiteralStringType :: StringType

_stringLiteralQuoteType :: QuoteType

_stringLiteralValue :: [PyChar]

_stringLiteralWhitespace :: [Whitespace]

StringLiteral

Fields

_stringLiteralAnn :: Ann a

_unsafeStringLiteralPrefix :: Maybe StringPrefix

_stringLiteralStringType :: StringType

_stringLiteralQuoteType :: QuoteType

_stringLiteralValue :: [PyChar]

_stringLiteralWhitespace :: [Whitespace]

RawBytesLiteral

Fields

_stringLiteralAnn :: Ann a

_unsafeRawBytesLiteralPrefix :: RawBytesPrefix

_stringLiteralStringType :: StringType

_stringLiteralQuoteType :: QuoteType

_stringLiteralValue :: [PyChar]

_stringLiteralWhitespace :: [Whitespace]

BytesLiteral

Fields

_stringLiteralAnn :: Ann a

_unsafeBytesLiteralPrefix :: BytesPrefix

_stringLiteralStringType :: StringType

_stringLiteralQuoteType :: QuoteType

_stringLiteralValue :: [PyChar]

_stringLiteralWhitespace :: [Whitespace]

Lenses

stringLiteralStringType :: forall a. Lens' (StringLiteral a) StringType

stringLiteralQuoteType :: forall a. Lens' (StringLiteral a) QuoteType

stringLiteralValue :: forall a. Lens' (StringLiteral a) [PyChar]

stringLiteralWhitespace :: forall a. Lens' (StringLiteral a) [Whitespace]

Rendering

showQuoteType :: QuoteType -> Char

showStringPrefix :: StringPrefix -> Text

showRawStringPrefix :: RawStringPrefix -> Text

showBytesPrefix :: BytesPrefix -> Text

showRawBytesPrefix :: RawBytesPrefix -> Text

Extra functions

isEscape :: PyChar -> Bool

Determine whether a PyChar is an escape character or not.