Language.Python.Syntax.Whitespace

Description

Whitespace

data Newline

A newline is either a carriage return, a line feed, or a carriage return followed by a line feed.

Constructors

CR

LF

CRLF

data Whitespace

Whitespace is either a space, a tab, a newline that continues the logical line (Continued), a newline that ends the logical line (Newline), or a Comment.

Despite not literally being whitespace, comments inside enclosed forms are treated as whitespace. See https://docs.python.org/3.5/reference/lexical_analysis.html#implicit-line-joining

Example and counterexample of comments as whitespace

( 1 +
  # here's a comment
 2 +
 3 # another comment
)
 

x = 5 + 5
# this line is not considered whitespace
y = x * 2

[ 1
, 2 # I'm whitespace
, 3
# also whitespace
]

Constructors

Space

Tab

Continued Newline [Whitespace]

Newline Newline

Comment (Comment ())

data Blank a

Lines which are "blank", meaning that they contain, if anything, only whitespace and/or a comment.

Constructors

Blank

Fields

_blankAnn :: Ann a

_blankWhitespaces :: [Whitespace]

_blankComment :: Maybe (Comment a)

class HasTrailingWhitespace s where

Every syntactic element contains the whitespace that immediately follows it.

This type class lets us access this trailing whitespace in many different types throughout hpython.

Minimal complete definition

trailingWhitespace

Methods

trailingWhitespace :: Lens' s [Whitespace]

class HasTrailingNewline (s :: [*] -> * -> *) where

A statement-containing thing may have a trailing newline

Some forms always have a trailing newline, which is why this class isn't just trailingNewline :: Lens' (s v a) (Maybe Newline)

Minimal complete definition

trailingNewline, setTrailingNewline

Methods

trailingNewline :: Traversal' (s v a) Newline

setTrailingNewline :: s v a -> Newline -> s v a

Indentation

data IndentLevel

Python has rules regarding the expansion of tabs into spaces and how to go about computing indentation after this is done.

See https://docs.python.org/3.5/reference/lexical_analysis.html#indentation

This data structure implements those rules as a monoid.

getIndentLevel :: IndentLevel -> Int

indentLevel :: Indent -> Int

absoluteIndentLevel :: Int -> Indent -> Int

newtype Indent

Constructors

MkIndent

Fields

unIndent :: FingerTree IndentLevel Whitespace

indentWhitespaces :: Iso' Indent [Whitespace]

An Indent is isomorphic to a list of Whitespace

indentIt :: [Whitespace] -> Indents a -> Indents a

Indent some indentation by a chunk

dedentIt :: Indents a -> Indents a

Deent some indentation by a chunk

data Indents a

A possibly annotated list of Indents.

Constructors

Indents

Fields

_indentsValue :: [Indent]

_indentsAnn :: Ann a

indentsValue :: forall a. Lens' (Indents a) [Indent]

indentsAnn :: forall a a. Lens (Indents a) (Indents a) (Ann a) (Ann a)

subtractStart :: Indents a -> Indents a -> Maybe (Indents a)

Subtract the first argument from the beginning of the second

Returns Nothing if the first list is not a prefix of the second.