Language.Python.Syntax

Documentation

type Raw f = f '[] ()

data Statement (v :: [*]) a

data Expr (v :: [*]) a

Parameters and arguments

Parameters

class HasParameters s where

Minimal complete definition

setParameters, parameters

Methods

setParameters :: [Raw Param] -> Raw s -> Raw s

parameters :: Lens' (Raw s) (CommaSep (Raw Param))

Positional

class HasPositional p v | p -> v where

Positional parameters/arguments

p_ :: Raw Expr -> Raw Arg

p_ :: Raw Ident -> Raw Param

Minimal complete definition

p_

Methods

p_ :: v -> p

Keyword

class HasKeyword p where

Keyword parameters/arguments

Minimal complete definition

k_

Methods

k_ :: Raw Ident -> Raw Expr -> p

data KeywordParam v a

Constructors

MkKeywordParam

Fields

_kpAnn :: a

_kpName :: Ident v a

_kpType :: Maybe ([Whitespace], Expr v a)

_kpEquals :: [Whitespace]

_kpExpr :: Expr v a

_KeywordParam :: Prism (Param v a) (Param '[] a) (KeywordParam v a) (KeywordParam '[] a)

Lenses

kpAnn :: forall v a. Lens' (KeywordParam v a) a

kpName :: forall v a. Lens' (KeywordParam v a) (Ident v a)

kpType :: forall v a. Lens' (KeywordParam v a) (Maybe ([Whitespace], Expr v a))

kpEquals :: forall v a. Lens' (KeywordParam v a) [Whitespace]

kpExpr :: forall v a. Lens' (KeywordParam v a) (Expr v a)

Decorators

decorated_ :: HasDecorators s => [Raw Expr] -> Raw s -> Raw s

class HasDecorators s where

Minimal complete definition

setDecorators, getDecorators, decorators

Methods

setDecorators :: [Raw Expr] -> Raw s -> Raw s

getDecorators :: Raw s -> [Raw Expr]

decorators :: Lens' (Raw s) [Raw Decorator]

Statements

Lines of code

blank_ :: Raw Line

Create a blank Line

class AsLine s where

Convert some data to a Line

Minimal complete definition

line_

Methods

line_ :: Raw s -> Raw Line

newtype Line v a

One or more lines of Python code

Constructors

Line

Fields

unLine :: Either ([Whitespace], Newline) (Statement v a)

Block bodies

class HasBody s where

forall ws s. id ~ getBody . flip (setBody ws) s

forall ws. let f x = setBody ws (getBody x) x in f ~ f . f

Minimal complete definition

setBody, getBody, body

Methods

setBody

Arguments

:: [Whitespace]
Indentation scheme for the new lines

-> NonEmpty (Raw Line)
Lines to become the new body

-> Raw s
Current code

-> Raw s

Set the block body of some code, first indententing the block by a specified amount, then indenting it by the level of its parent

For example, if we wanted to set the body of function_b to

while True:
  pass

in

def function_a():
  def function_b():
    pass
    if True:
      pass
    else:
      pass

then we would get

def function_a():
  def function_b():
    while True:
      pass

getBody :: Raw s -> NonEmpty (Raw Line)

View the block body of some code without the indentation from its outer context

For example, if we ran getBody function_b in this code:

def function_a():
  def function_b():
    pass
    if True:
      pass
    else:
      pass

then we would get

pass
if True:
  pass
else:
  pass

body :: Lens' (Raw s) (Raw Suite)

modifyBody

Arguments

:: HasBody s

=> [Whitespace]
New indentation scheme

-> (NonEmpty (Raw Line) -> NonEmpty (Raw Line))
Modification function

-> Raw s
Existing code

-> Raw s

Modify the block body of some code

Function definitions

def_ :: Raw Ident -> [Raw Param] -> NonEmpty (Raw Line) -> Raw Statement

data Fundef v a

Constructors

MkFundef

Fields

_fdAnn :: a

_fdDecorators :: [Decorator v a]

_fdIndents :: Indents a

_fdAsync :: Maybe (NonEmpty Whitespace)

_fdDefSpaces :: NonEmpty Whitespace

_fdName :: Ident v a

_fdLeftParenSpaces :: [Whitespace]

_fdParameters :: CommaSep (Param v a)

_fdRightParenSpaces :: [Whitespace]

_fdReturnType :: Maybe ([Whitespace], Expr v a)

_fdBody :: Suite v a

_Fundef :: Prism (Statement v a) (Statement '[] a) (Fundef v a) (Fundef '[] a)

mkFundef :: Raw Ident -> NonEmpty (Raw Line) -> Raw Fundef

Create a minimal valid function definition

Lenses

fdAnn :: forall v a. Lens' (Fundef v a) a

fdDecorators :: forall v a. Lens' (Fundef v a) [Decorator v a]

fdIndents :: forall v a. Lens' (Fundef v a) (Indents a)

fdAsync :: forall v a. Lens' (Fundef v a) (Maybe (NonEmpty Whitespace))

fdDefSpaces :: forall v a. Lens' (Fundef v a) (NonEmpty Whitespace)

fdName :: forall v a. Lens' (Fundef v a) (Ident v a)

fdLeftParenSpaces :: forall v a. Lens' (Fundef v a) [Whitespace]

fdParameters :: forall v a. Lens' (Fundef v a) (CommaSep (Param v a))

fdRightParenSpaces :: forall v a. Lens' (Fundef v a) [Whitespace]

fdReturnType :: forall v a. Lens' (Fundef v a) (Maybe ([Whitespace], Expr v a))

fdBody :: forall v a. Lens' (Fundef v a) (Suite v a)

Assignment

(.=) :: Raw Expr -> Raw Expr -> Raw Statement

Flow control

Else clauses

else_ :: HasElse s => NonEmpty (Raw Line) -> Raw s -> Raw s

class HasElse s where

Minimal complete definition

getElse, setElse

Methods

getElse :: Raw s -> Maybe (Raw Else)

setElse :: [Whitespace] -> Maybe (Raw Else) -> Raw s -> Raw s

Break

break_ :: Raw Statement

For loops

for_ :: Raw Expr -> Raw Expr -> NonEmpty (Raw Line) -> Raw Statement

data For v a

Constructors

MkFor

Fields

_forAnn :: a

_forIndents :: Indents a

_forAsync :: Maybe (NonEmpty Whitespace)

_forFor :: [Whitespace]

_forBinder :: Expr v a

_forIn :: [Whitespace]

_forCollection :: Expr v a

_forBody :: Suite v a

_forElse :: Maybe (Else v a)

_For :: Prism (Statement v a) (Statement '[] a) (For v a) (For '[] a)

mkFor :: Raw Expr -> Raw Expr -> NonEmpty (Raw Line) -> Raw For

If statements

if_ :: Raw Expr -> NonEmpty (Raw Line) -> Raw If

elif_ :: Raw Expr -> NonEmpty (Raw Line) -> Raw If -> Raw If

data If v a

Constructors

MkIf

Fields

_ifAnn :: a

_ifIndents :: Indents a

_ifIf :: [Whitespace]

_ifCond :: Expr v a

_ifBody :: Suite v a

_ifElifs :: [Elif v a]

_ifElse :: Maybe (Else v a)

_If :: Prism (Statement v a) (Statement '[] a) (If v a) (If '[] a)

mkIf :: Raw Expr -> NonEmpty (Raw Line) -> Raw If

Create a minimal valid If

data Elif v a

Constructors

MkElif

Fields

_elifIndents :: Indents a

_elifElif :: [Whitespace]

_elifCond :: Expr v a

_elifBody :: Suite v a

_Elif :: Iso' (Elif v a) (Indents a, [Whitespace], Expr v a, Suite v a)

mkElif :: Raw Expr -> NonEmpty (Raw Line) -> Raw Elif

Create a minimal valid Elif

data Else v a

Constructors

MkElse

Fields

_elseIndents :: Indents a

_elseElse :: [Whitespace]

_elseBody :: Suite v a

_Else :: Iso' (Else v a) (Indents a, [Whitespace], Suite v a)

mkElse :: NonEmpty (Raw Line) -> Raw Else

Create a minimal valid Else

Lenses

ifAnn :: forall v a. Lens' (If v a) a

ifIndents :: forall v a. Lens' (If v a) (Indents a)

ifIf :: forall v a. Lens' (If v a) [Whitespace]

ifCond :: forall v a. Lens' (If v a) (Expr v a)

ifBody :: forall v a. Lens' (If v a) (Suite v a)

ifElifs :: forall v a. Lens' (If v a) [Elif v a]

ifElse :: forall v a. Lens' (If v a) (Maybe (Else v a))

elifIndents :: forall v a. Lens' (Elif v a) (Indents a)

elifElif :: forall v a. Lens' (Elif v a) [Whitespace]

elifCond :: forall v a. Lens' (Elif v a) (Expr v a)

elifBody :: forall v a. Lens' (Elif v a) (Suite v a)

elseIndents :: forall v a. Lens' (Else v a) (Indents a)

elseElse :: forall v a. Lens' (Else v a) [Whitespace]

elseBody :: forall v a v. Lens (Else v a) (Else v a) (Suite v a) (Suite v a)

Pass

pass_ :: Raw Statement

Return

return_ :: Raw Expr -> Raw Statement

While loops

while_ :: Raw Expr -> NonEmpty (Raw Line) -> Raw Statement

data While v a

Constructors

MkWhile

Fields

_whileAnn :: a

_whileIndents :: Indents a

_whileWhile :: [Whitespace]

_whileCond :: Expr v a

_whileBody :: Suite v a

_While :: Prism (Statement v a) (Statement '[] a) (While v a) (While '[] a)

mkWhile :: Raw Expr -> NonEmpty (Raw Line) -> Raw While

Create a minimal valid While

Expressions

expr_ :: Raw Expr -> Raw Statement

var_ :: String -> Raw Expr

Function calls

call_ :: Raw Expr -> [Raw Arg] -> Raw Expr

data Call v a

Constructors

MkCall

Fields

_callAnn :: a

_callFunction :: Expr v a

_callLeftParen :: [Whitespace]

_callArguments :: Maybe (CommaSep1' (Arg v a))

_callRightParen :: [Whitespace]

_Call :: Prism (Expr v a) (Expr '[] a) (Call v a) (Call '[] a)

mkCall :: Raw Expr -> Raw Call

Create a minimal valid Call

Lenses

callAnn :: forall v a. Lens' (Call v a) a

callFunction :: forall v a. Lens' (Call v a) (Expr v a)

callLeftParen :: forall v a. Lens' (Call v a) [Whitespace]

callArguments :: forall v a. Lens' (Call v a) (Maybe (CommaSep1' (Arg v a)))

callRightParen :: forall v a. Lens' (Call v a) [Whitespace]

Literals

list_ :: [Raw Expr] -> Raw Expr

none_ :: Raw Expr

str_ :: String -> Raw Expr

true_ :: Raw Expr

false_ :: Raw Expr

Dereferencing

(/>) :: Raw Expr -> Raw Ident -> Raw Expr infixl 9

Binary operators

is_ :: Raw Expr -> Raw Expr -> Raw Expr infixl 1

(.*) :: Raw Expr -> Raw Expr -> Raw Expr infixl 7

(.-) :: Raw Expr -> Raw Expr -> Raw Expr infixl 6

(.+) :: Raw Expr -> Raw Expr -> Raw Expr infixl 6

(.==) :: Raw Expr -> Raw Expr -> Raw Expr infixl 1