Language.Python.Syntax.Strings

Description

Python string literals.

See https://docs.python.org/3.5/reference/lexical_analysis.html#string-and-bytes-literals

Documentation

data QuoteType

Double or single quotation marks?

"Double quotes"
"""Double quotes"""
'Single quotes'
'''Single quotes'''

Constructors

SingleQuote

DoubleQuote

data StringType

Three paris of quotations or one?

"""Long string"""
'''Also long string'''
"Short string"
'Also short string'

Constructors

ShortString

LongString

data StringPrefix

In Python 3.5, a prefix of u or U is allowed, but doesn't have any meaning. They exist for backwards compatibility with Python 2.

See https://www.python.org/dev/peps/pep-0414/

Constructors

Prefix_u

Prefix_U

data RawStringPrefix

Raw strings are prefixed with either r or R.

Constructors

Prefix_r

Prefix_R

data BytesPrefix

This prefix indicates it's a bytes literal rather than a string literal.

Constructors

Prefix_b

Prefix_B

data RawBytesPrefix

A string of raw bytes can be indicated by a number of prefixes

Constructors

Prefix_br

Prefix_Br

Prefix_bR

Prefix_BR

Prefix_rb

Prefix_rB

Prefix_Rb

Prefix_RB

hasStringPrefix :: StringLiteral a -> Bool

Most types of StringLiteral have prefixes. Plain old strings may have an optional prefix, but it is meaningless.

data StringLiteral a

A StringLiteral, complete with a prefix, information about quote type and number, and a list of PyChars.

Like many other data types in hpython, it has an annotation and trailing whitespace.

Constructors

RawStringLiteral

Fields

_stringLiteralAnn :: a

_unsafeRawStringLiteralPrefix :: RawStringPrefix

_stringLiteralStringType :: StringType

_stringLiteralQuoteType :: QuoteType

_stringLiteralValue :: [PyChar]

_stringLiteralWhitespace :: [Whitespace]

StringLiteral

Fields

_stringLiteralAnn :: a

_unsafeStringLiteralPrefix :: Maybe StringPrefix

_stringLiteralStringType :: StringType

_stringLiteralQuoteType :: QuoteType

_stringLiteralValue :: [PyChar]

_stringLiteralWhitespace :: [Whitespace]

RawBytesLiteral

Fields

_stringLiteralAnn :: a

_unsafeRawBytesLiteralPrefix :: RawBytesPrefix

_stringLiteralStringType :: StringType

_stringLiteralQuoteType :: QuoteType

_stringLiteralValue :: [PyChar]

_stringLiteralWhitespace :: [Whitespace]

BytesLiteral

Fields

_stringLiteralAnn :: a

_unsafeBytesLiteralPrefix :: BytesPrefix

_stringLiteralStringType :: StringType

_stringLiteralQuoteType :: QuoteType

_stringLiteralValue :: [PyChar]

_stringLiteralWhitespace :: [Whitespace]

data PyChar

A character in a string literal. This is a large sum type, with a catch-all of a Haskell Char.

Constructors

Char_newline

Char_octal1 OctDigit

Char_octal2 OctDigit OctDigit

Char_octal3 OctDigit OctDigit OctDigit

Char_hex HeXDigit HeXDigit

Char_uni16 HeXDigit HeXDigit HeXDigit HeXDigit

Char_uni32 HeXDigit HeXDigit HeXDigit HeXDigit HeXDigit HeXDigit HeXDigit HeXDigit

Char_esc_bslash

Char_esc_singlequote

Char_esc_doublequote

Char_esc_a

Char_esc_b

Char_esc_f

Char_esc_n

Char_esc_r

Char_esc_t

Char_esc_v

Char_lit Char

isEscape :: PyChar -> Bool

Determine whether a PyChar is an escape character or not.

fromHaskellString :: String -> [PyChar]

Convert a Haskell string to a list of PyChar. This is useful when writing Python in Haskell.

showStringPrefix :: StringPrefix -> Text

showRawStringPrefix :: RawStringPrefix -> Text

showBytesPrefix :: BytesPrefix -> Text

showRawBytesPrefix :: RawBytesPrefix -> Text

showQuoteType :: QuoteType -> Char

stringLiteralValue :: forall a. Lens' (StringLiteral a) [PyChar]